%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addNewLine.RTF}
\alias{addNewLine.RTF}
\alias{RTF.addNewLine}
\alias{addNewLine,RTF-method}

 \title{Insert a new line into the RTF document}

 \description{
	Insert a new line into the RTF document.
 }

 \usage{
\method{addNewLine}{RTF}(this, n=1, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{n}{Number of lines to add. Default is 1. \bold{optional}}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
