%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addParagraph.RTF}
\alias{addParagraph.RTF}
\alias{RTF.addParagraph}
\alias{addParagraph.RTF}
\alias{addParagraph,RTF-method}


 \title{Insert a paragraph into the RTF document}

 \description{
	Insert a paragraph into the RTF document.
 }

 \usage{\method{addParagraph}{RTF}(this, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{A character \code{\link[base]{vector}} of text to add.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
