%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addPng.RTF}
\alias{addPng.RTF}
\alias{RTF.addPng}
\alias{addPng,RTF-method}

 \title{Insert an existing PNG image into the RTF document}

 \description{
	Insert an existing PNG image into the RTF document.
 }

 \usage{
\method{addPng}{RTF}(this, file, width=3, height=0.3, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{file}{Image file path.}
 	\item{width}{Plot output width in inches.}
 	\item{height}{Plot output height in inches.}
 	\item{...}{Not used.}
 }

 \details{
 	Add existing PNG file to RTF document.
 }

 \author{Michael Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
