%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addText.RTF}
\alias{addText.RTF}
\alias{RTF.addText}
\alias{addText,RTF-method}

 \title{Insert text into the RTF document}

 \description{
	Insert text into the RTF document.
 }

 \usage{
\method{addText}{RTF}(this, ..., bold=FALSE, italic=FALSE)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{bold}{Bold text. \bold{optional}.}
 	\item{italic}{Italic text. \bold{optional}.}
 	\item{...}{Any number of character strings to concatenate.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
