%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addTOC.RTF}
\alias{addTOC.RTF}
\alias{RTF.addTOC}
\alias{addTOC,RTF-method}

 \title{Insert table of contents field}

 \description{
	Insert table of contents field.
 }

 \usage{
\method{addTOC}{RTF}(this, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
