% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLog10ForTakeLog10.R
\name{applyLog10ForTakeLog10}
\alias{applyLog10ForTakeLog10}
\title{Take Log10(x) or 10^(x) for a vector x}
\usage{
applyLog10ForTakeLog10(
  x = c(),
  takeLog10 = c(),
  reverse = FALSE,
  calcGradient = FALSE
)
}
\arguments{
\item{x}{Named vector with parameter values}

\item{takeLog10}{Named vector of booleans indicating if log10 is/should be 
taken of the model parameters.}

\item{reverse}{Boolean indicating if 10^ should be taken instead of log10() 
(Default: FALSE).}

\item{calcGradient}{Boolean indicating if gradient should be calculated
(Default: FALSE)}
}
\value{
Named vector where the parameters, which are
}
\description{
Take Log10(x) or 10^(x) for a vector x
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'timeDependent')
initialGuess.vec <- optimObject.orig[["initialGuess.vec"]]
lb.vec <- optimObject.orig[["lb.vec"]]
takeLog10 <- optimObject.orig[["takeLog10"]]
applyLog10ForTakeLog10(lb.vec, takeLog10) 


x <- 1:3
names(x) <- c("A","B","C")
takeLog10 <- c(TRUE, TRUE, FALSE)
names(takeLog10) <- names(x)
applyLog10ForTakeLog10(x,takeLog10 = takeLog10)
applyLog10ForTakeLog10(x,takeLog10 = takeLog10, calcGradient = TRUE)
}
