%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{decreaseIndent.RTF}
\alias{decreaseIndent.RTF}
\alias{RTF.decreaseIndent}
\alias{decreaseIndent.RTF}
\alias{decreaseIndent,RTF-method}


 \title{Decrease RTF document indent}

 \description{
	Decrease RTF document indent.
 }

 \usage{\method{decreaseIndent}{RTF}(this, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }
\keyword{internal}
\keyword{methods}
