%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{done.RTF}
\alias{done.RTF}
\alias{RTF.done}
\alias{done,RTF-method}

 \title{Write and close the RTF output}

 \description{
	Write and close the RTF output.
 }

 \usage{
\method{done}{RTF}(this, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael Schaffer}

 \seealso{
 	For more information see \code{\link{RTF}}.
 }


\keyword{methods}
