%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{endParagraph.RTF}
\alias{endParagraph.RTF}
\alias{RTF.endParagraph}
\alias{endParagraph.RTF}
\alias{endParagraph,RTF-method}


 \title{End a paragraph in the RTF document}

 \description{
	End a paragraph in the RTF document.
 }

 \usage{\method{endParagraph}{RTF}(this, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
