% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingHelper.R
\name{fittingHelper}
\alias{fittingHelper}
\title{Get best fit for initial guesses}
\usage{
fittingHelper(optimObject, plot = TRUE, nInitialGuesses = 50, titlePrefix = "")
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted').
optimObject$fixed[["signum_TF"]] has to be set to 1 or -1}

\item{plot}{Boolean value indicating if fitting results should be plotted and saved as file.
Plots include for each fit plot of the components of the transient function
and the position of the fit in the waterfall plot based on the value of the
optimization measure the fit reached. For the best fit the components of
 the transient function are plotted as well as the parameter distribution
 over all fits and a water fall plot of all fits.}

\item{nInitialGuesses}{Number of initial guesses}

\item{titlePrefix}{Prefix of file names of plots, if plot  was set to TRUE}
}
\value{
List with the parameters of the best fit ('res.pars'),
the optimization measure value of the best fit ('value'),
and the ggplot object of the plot of the best fit ('bestFit.plot')
}
\description{
Get best fit for predefined number of initial guesses
}
\examples{
data <- getExampleDf()
data <- scaleTimeCol(data)
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
optim.res <- fittingHelper(optimObject.orig, plot = TRUE,
                            titlePrefix = "fullModel_plus_")
}
