% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFittingResult.R
\name{getBestFittingResult}
\alias{getBestFittingResult}
\title{Fit function for positive OR negative sign}
\usage{
getBestFittingResult(
  optimObject,
  parStr = "signum_TF",
  parVal = 1,
  plot,
  titlePrefix = ""
)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame
with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of
lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and
ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted').}

\item{parVal}{Value of variable defined in parString  (Default: 1)}

\item{plot}{Boolean value indicating if fitting results should be plotted
and saved as file}

\item{titlePrefix}{Prefix of file names of plots, if plot  was set
to TRUE}

\item{parString}{String of variable based on which comparison is conducted (Default: "signum_TF")}
}
\value{
optimObject, to which fitting result of the best fit ('fitted'),
optimization measure value of this fit ('value'), and, if plot=TRUE, the
plot to this fit ('bestFit.plot') has been added.
}
\description{
Fits function for positive OR negative sign (signum_TF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
res.all.plus <- getBestFittingResult(
        optimObject.orig, parStr = "signum_TF", parVal = 1, plot = TRUE,
        titlePrefix = "fullModel_plus_")
}
