% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDoseResponseExampleDf.R
\name{getDoseResponseExampleDf}
\alias{getDoseResponseExampleDf}
\title{Generate example data frame for defined dose-response RTF parameters}
\usage{
getDoseResponseExampleDf()
}
\value{
Dataframe with data following an RTF for different doses d
 with predefined parameters plus some added noise.
}
\description{
Generates an example dataframe with data following an RTF with
predefined parameters plus some added noise, for different doses d.
}
\examples{
getDoseResponseExampleDf()
}
