% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDynamic2DPlot.R
\name{getDynamic2DPlot}
\alias{getDynamic2DPlot}
\title{Generates dynamic 2D plot based on the RTF parameters for multiple time 
series}
\usage{
getDynamic2DPlot(
  dim1Vec,
  dim2Vec,
  ID,
  species,
  res.lst,
  metaInfo = c(),
  positionLeft = 0,
  positionTop = 0
)
}
\arguments{
\item{dim1Vec}{Vector with values of first dimension.}

\item{dim2Vec}{Vector with values of second dimension. Has same length of 
dim1Vec}

\item{ID}{Vector with unique identifiers with same length as dim1Vec and 
dim2Vec.}

\item{species}{Vector with names of molecular species with same length as 
dim1Vec and dim2Vec.}

\item{res.lst}{List with the RTF result for each time series}

\item{metaInfo}{(Optional) Vector specifying  meta information for
each time series. Needs to have the same length as the number of rows of 
param.df.}

\item{positionLeft}{Integer specifying the horizontal position of the subplot 
(Default: 0)}

\item{positionTop}{Integer specifying the vertical position of the subplot 
(Default: 0)}
}
\value{
Plotly object of the dynamic 2D plot.
}
\description{
Generates dynamic 2D plot based on the RTF parameters for 
multiple time series, where each point corresponds to a single time series. 
By hovering over a point the corresponding time-resolved behavior is 
displayed in an additional smaller subplot.
}
\examples{
\dontrun{
data(almadenTimeSeries)
data(almadenParams)
data(almadenResLst)
timeSeries <- almadenTimeSeries # first column needs to be "time"
colNames <- colnames(timeSeries[2:ncol(timeSeries)])
species <- sub("_[^_]+$", "", colNames)
conditionID <- gsub(".*_", "", colNames)

# # Besides returning the fitted parameters, this function also generates 
# # a RDS with the final models, which is needed later on
# fileString <- "testFile"
# param.df <- getParamsFromMultipleTimeSeries(
#   almadenTimeSeries,
#   fileString = fileString,
#   saveFolderPath = tempdir(),
#   nInitialGuesses = 100
# )
# res.lst <- readRDS(file = paste0(tempdir(), "/", fileString, ".RDS"))

numClust <- NULL
plots <- getLowDimensionalRTFPlots(
  df = almadenParams,
  metaInfo = species, 
  metaInfoName = "species",
  maxTime = max(timeSeries$time),
  numClust = numClust
)
    
df.umapData <- plots[["umap.data"]]
clustID <- df.umapData$clustID
ID <- row.names(df.umapData)
plt <- getDynamic2DPlot(dim1Vec = df.umapData$UMAP1,
                        dim2Vec = df.umapData$UMAP2,
                        ID = ID,
                        species = species,
                        res.lst = almadenResLst,
                        metaInfo = conditionID # alternatively: clustID
)
}
}
