% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpData.R
\name{getExpData}
\alias{getExpData}
\title{Generate data frame from experimental data for RTF}
\usage{
getExpData(file = "", tCol = "", quantCols = c(), sdExpCol = "")
}
\arguments{
\item{file}{File name of experimental data. Allowed formats:
.csv, .xlsx, .xls. file should at least contain two columns, which if not
otherwise specified will be named 't' and 'y'. Alternatively, the column
corresponding to time and the columns containing quantitative information can
be specified using tCol and quantCols, respectively.}

\item{tCol}{If file is provided, tCol can be used to specify column
indicating time}

\item{quantCols}{If file is provided, quantCols can be used to specify
columns containing quantitative information. This can be used, e.g., where
the columns in quantCols correspond to replicates at the same time point.}

\item{sdExpCol}{If file is provided, sdExpCol can be used to explicitly
define the column with the standard deviation of the experimental data. If
not provided the standard deviation will be estimated during fitting.}
}
\value{
Data frame containing columns named 't' (time) and
'y' (quantitative value). If .csv, .xlsx, or .xls file name is provided for
parameter 'file' is provided, this will be used, if no file name is
provided data will be simulated.
}
\description{
Generate data frame containing all columns needed for RTF
}
\examples{
data <- getExpData()
# data <- getExpData(file="ExampleData.xls", tCol="time",
# quantCols=c("Replikat1", "Replikat2", "Replikat3"))
}
