% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFittingResult.R
\name{getFittingResult}
\alias{getFittingResult}
\title{Fit function for positive AND negative sign}
\usage{
getFittingResult(optimObject, plot = TRUE, titlePrefixPrefix = "")
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame
with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds
('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec
('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted')}

\item{plot}{Boolean value indicating if fitting results should be plotted
and saved as file}

\item{titlePrefixPrefix}{Prefix of file names of plots, if plot  was set
to TRUE}
}
\value{
List of two optimObjects for a positive and a negative signum_TF
('plus1' and 'minus1'). Each including fitting result of the best
fit ('fitted'), optimization measure value of this fit ('value'), and, if
plot=TRUE, the plot to this fit ('bestFit.plot') has been added.
}
\description{
Fits function for positive AND negative sign (signum_TF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
            data, modus = 'RetardedTransientDynamics')
res.all.plusMinus <- getFittingResult(
            optimObject.orig, plot = TRUE, titlePrefixPrefix = "fullModel_")
}
