% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialGuessResults.R
\name{getInitialGuessResults}
\alias{getInitialGuessResults}
\title{Get fits for defined number of initial guesses}
\usage{
getInitialGuessResults(optimObject, objFunct, nInitialGuesses, plot = TRUE)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted').
optimObject$fixed[["signum_TF"]] has to be set to 1 or -1}

\item{objFunct}{Name of the objective function}

\item{nInitialGuesses}{Integer corresponding to number of initial guesses}

\item{plot}{Boolean value indicating if fitting results should be plotted and saved as file.
Plots include for each fit plot of the components of the transient function
and the position of the fit in the waterfall plot based on the value of the
optimization measure the fit reached. For the best fit the components of
 the transient function are plotted as well as the parameter distribution
 over all fits and a water fall plot of all fits.}
}
\value{
List of with optimization results and plots to each fit ('res.lst')
and optimization result and plot to best fit ('final'), both with optional plots
}
\description{
Get fits for defined number of initial guesses and add optional
plots (incl. for best fit ('final') waterfall plot and for all parameters
histograms of the parameter values across all fits)
}
\examples{
data <- getExampleDf()
data <- scaleTimeCol(data)
optimObject.orig <- initializeOptimObject(
                data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
optim.res <- getInitialGuessResults(
                optimObject.orig, objFunct, nInitialGuesses, plot = TRUE)
}
