% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialGuessVec.R
\name{getInitialGuessVec}
\alias{getInitialGuessVec}
\title{Get list of initial guesses}
\usage{
getInitialGuessVec(initialGuess.vec, lb.vec, ub.vec, nInitialGuesses)
}
\arguments{
\item{initialGuess.vec}{Vector of the default initial guesses for each
model parameter}

\item{lb.vec}{Vector of the lower bounds for each model parameter}

\item{ub.vec}{Vector of the upper bounds for each model parameter}

\item{nInitialGuesses}{Integer corresponding to number of initial guesses}
}
\value{
List of default initial guess for each model parameter
a defined number of further initial guesses lying in between the lower and
upper bound of each model parameter
}
\description{
Get list of default initial guess for each model parameter and
a defined number of further initial guesses lying in between the lower and
upper bound of each model parameter
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
initialGuess.vec <- optimObject.orig[["initialGuess.vec"]]
lb.vec <- optimObject.orig[["lb.vec"]]
ub.vec <- optimObject.orig[["ub.vec"]]
nInitialGuesses <- 50
initialGuess.vec.lst <- getInitialGuessVec(
                          initialGuess.vec, lb.vec, ub.vec, nInitialGuesses)
}
