% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialGuessVecLst.R
\name{getInitialGuessVecLst}
\alias{getInitialGuessVecLst}
\title{Get list of initial guesses}
\usage{
getInitialGuessVecLst(
  initialGuess.vec,
  lb.vec,
  ub.vec,
  takeLog10,
  nInitialGuesses = 50
)
}
\arguments{
\item{initialGuess.vec}{Vector of the default initial guesses for each
model parameter}

\item{lb.vec}{Vector of the lower bounds for each model parameter}

\item{ub.vec}{Vector of the upper bounds for each model parameter}

\item{takeLog10}{Vector of booleans indicating if log10 is/should be taken
of the model parameters.}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) (Default: 50).}
}
\value{
List of the default initial guess for each model parameter and
a defined number of further initial guesses lying between the lower and
upper bound of each model parameter
}
\description{
Get list of the default initial guess for each model parameter 
and a defined number of further initial guesses lying between the lower 
and upper bound of each model parameter
}
\examples{
data <- getSimData()
optimObject.orig <- initializeOptimObject(data, modus = 'singleDose')
initialGuess.vec <- optimObject.orig[["initialGuess.vec"]]
lb.vec <- optimObject.orig[["lb.vec"]]
ub.vec <- optimObject.orig[["ub.vec"]]
takeLog10 <- optimObject.orig[["takeLog10"]]
initialGuess.vec.lst <- getInitialGuessVecLst(
                          initialGuess.vec, lb.vec, ub.vec, takeLog10)
}
