% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLowDimensionalRTFPlots.R
\name{getLowDimensionalRTFPlots}
\alias{getLowDimensionalRTFPlots}
\title{Generate plots of low dimensional representation of RTF parameters}
\usage{
getLowDimensionalRTFPlots(df, metaInfo, metaInfoName, maxTime = 10)
}
\arguments{
\item{df}{Data frame with the RTF parameter values. 
Its columns 'alpha', 'gamma', 'A', 'B', 'b', 'tau', and 'signum_TF'
represent the different RTF parameters. The rows correspond to the different
time series.}

\item{metaInfo}{String of the column name with meta information (.e.g.
groups.}

\item{metaInfoName}{String of the name of the meta information.}

\item{maxTime}{Time point up to which the dynamics should be plotted.}
}
\value{
List containing UMAP plot colored by meta info ('umap.metaInfo'), 
UMAP plot colored by cluster ID ('umap.cluster'), plot of parameter 
distribution ('parDistribution'), unscaled dynamics plots 
('dynamics.notscaled'), scaled dynamics plots ('dynamics.scaled'), 
UMAP result data frame ('umap.data'), and data frame with the dynamics values 
('dynamics.data')
}
\description{
Generates plots of low dimensional representation of RTF 
parameters
}
\examples{
data(strasenParams)
metaInfo <- sub("_[^_]+$", "", row.names(strasenParams))
metaInfoName <- "species"
res <- getLowDimensionalRTFPlots(df = strasenParams,
                                metaInfo = metaInfo,
                                metaInfoName = metaInfoName,
                                maxTime = 10)
}
