% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultiStartResults.R
\name{getMultiStartResults}
\alias{getMultiStartResults}
\title{Get fits for defined number of initial guesses}
\usage{
getMultiStartResults(optimObject, objFunct, nInitialGuesses)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted').
optimObject$fixed[["signum_TF"]] has to be set to 1 or -1}

\item{objFunct}{Name of the objective function}

\item{nInitialGuesses}{Integer corresponding to number of initial guesses}
}
\value{
optimObject supplemented by all optimization results ('optimResults') and
the ebst optimization result ('bestOptimResult')
}
\description{
Get fits for defined number of initial guesses and add optional
plots (incl. for best fit ('final') waterfall plot and for all parameters
histograms of the parameter values across all fits)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
                data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
optim.res <- getMultiStartResults(
                optimObject.orig, objFunct, nInitialGuesses)
}
