% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamsFromMultipleTimeCourses.R
\name{getParamsFromMultipleTimeCourses}
\alias{getParamsFromMultipleTimeCourses}
\title{Get RTF parameters for multiple time courses}
\usage{
getParamsFromMultipleTimeCourses(
  df,
  doRTFPlots = TRUE,
  fileString = "lowDimRTF",
  readInParamRdsFilePath = "",
  saveFolderPath = "",
  modelReduction = FALSE,
  nInitialGuesses = 50,
  sameSign = TRUE,
  plotFitsToSingleFile = TRUE,
  plotFitOnly = FALSE,
  plotAllPointsWaterfall = FALSE
)
}
\arguments{
\item{df}{Data frame in which the first column must contain the measurement 
time points, while all subsequent columns correspond to the different time 
courses.}

\item{doRTFPlots}{Boolean indicating if RTF plots should be generated.}

\item{fileString}{String the name of the created .rds file with the RTF
parameters should contain. If this string has length 0 or if
readInParamRdsFilePath is defined no .rds file will be created.}

\item{readInParamRdsFilePath}{If the RTF parameters to time courses have
already been determined previously, the path of the resulting .rds file can
be specified  using readInParamRdsFilePath, so that the time-intensive
calculation of the RTF parameter does not have to be repreated a second time.}

\item{saveFolderPath}{Path of folder to where plots and rds files should be
saved. Default: current folder.}

\item{modelReduction}{Boolean indicating if model reduction should be
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses
(in addition to the default initial guess) used for all four combinations
of signSus and signTrans, which both can be -1 or 1 (Default: 50).}

\item{sameSign}{Boolean indicating if sign of sustained RTF part
(signSus) and transient RTF part (signTrans) should be equal
(Default: TRUE) (Only relevant if readInParamRdsFilePath is not provided).}

\item{plotFitsToSingleFile}{Boolean indicating if plots should be returned
as a single file.}

\item{plotFitOnly}{Plot fit only without additional information as provided
using function plotRTF().}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be
plotted in waterfall plot (Default: FALSE).
If FALSE, all values up to the median of those values are plotted.}
}
\value{
List of data frame containing the RTF parameters for each time
course, where the rows correspond to the different time courses and
the columns to the RTF parameters ('param.df'), and list with the RTF result
for each time course ('RTFmodelLst').
}
\description{
Get RTF parameters for multiple time courses specified in
data frame.
}
\examples{
data(strasen)
df <- strasen[, 1:3]
param.df <- getParamsFromMultipleTimeCourses(df)
}
