% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositiveParNames.R
\name{getPositiveParNames}
\alias{getPositiveParNames}
\title{Get names of parameters having no negative values}
\usage{
getPositiveParNames(lb.vec, ub.vec, initialGuess.vec)
}
\arguments{
\item{lb.vec}{Named vector of lower bound for each model parameter}

\item{ub.vec}{Named vector of upper bound for each model parameter}

\item{initialGuess.vec}{Named vector of initial guesses for each model 
parameter}
}
\value{
Vector of parameters having no negative values in
initialGuess.vec, lb.vec, and ub.vec
}
\description{
Returns the names of parameters having no negative values in
initialGuess.vec, lb.vec, and ub.vec.
}
\examples{
len <- 20
lb.vec <- rnorm(len)
ub.vec <- lb.vec + runif(len, 0, 2)
initialGuess.vec <- runif(len, lb.vec, ub.vec)
names(lb.vec) <- names(ub.vec) <- names(initialGuess.vec) <-
                                                      paste0("param", 1:len)
getPositiveParNames(lb.vec, ub.vec, initialGuess.vec)
}
