% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTFResDf.R
\name{getRTFResDf}
\alias{getRTFResDf}
\title{Generate a data frame with RTF results for defined time points (and doses)}
\usage{
getRTFResDf(par, modus = "singleDose", t, doses = c(), scale = TRUE)
}
\arguments{
\item{par}{Named vector of the parameter values used for RTF (and, in the 
case of the dose-dependent RTF, for the Hill equations).}

\item{modus}{Modus ('singleDose' or 'doseDependent')}

\item{t}{Vector of time points}

\item{doses}{Vector of doses. Only relevant if modus = 'doseDependent'.}

\item{scale}{Boolean, indicates if time-dependent parameters t, tau,
alpha, beta, and gamma should be scaled (Default: TRUE).}
}
\value{
Data frame with column for defined time points ('t') and the 
predicted quantitative values resulting from calculating RTF with the 
specified parameters fir these time points ('y'). If modus = 'doseDependent' 
there is an additional column for the dose ('d').
}
\description{
Generate a data frame with RTF results for defined time 
points (and doses).
}
\examples{
rtfPar <- c(
    alpha = 1.00, beta = 1.00, gamma = 1.00, A = 1.05,
    B = 3.05, b = -0.28, tau = -1, signSus = 1, signTrans = 1
)
t <- c(
    0, 0.71, 1.42, 2.14, 2.85, 3.57, 4.28, 5, 5.71, 6.42,
    7.14, 7.85, 8.57, 9.28, 10
)
df <- getRTFResDf(par = rtfPar, t = t)
}
