% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignalSusPlusOffset.R
\name{getSignalSusPlusOffset}
\alias{getSignalSusPlusOffset}
\title{Calculate sum of the sustained part of the RTF and offset b}
\usage{
getSignalSusPlusOffset(alphaInv, A, b, tau, signum_TF, t)
}
\arguments{
\item{alphaInv}{alphaInv}

\item{A}{A}

\item{b}{b}

\item{tau}{tau}

\item{signum_TF}{signum_TF}

\item{t}{Time point}
}
\value{
Sum of the sustained part of the RTF and offset b for a defined
time point t.
}
\description{
Calculate sum of the sustained part of the RTF and offset b
for a defined time point.
}
\examples{
getSignalSusPlusOffset(
        alphaInv = 1.00,
        A = 1.05,
        b = -0.28, tau = -1, signum_TF = 1,
        t = 0.6)
}
