% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignalTransPlusOffset.R
\name{getSignalTransPlusOffset}
\alias{getSignalTransPlusOffset}
\title{Calculate sum of the transient part of the RTF and offset p_0}
\usage{
getSignalTransPlusOffset(
  tau_1,
  tau_2,
  A_trans,
  p_0,
  T_shift,
  signum_TF,
  t_prime
)
}
\arguments{
\item{tau_1}{tau_1}

\item{tau_2}{tau_2}

\item{A_trans}{A_trans}

\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{signum_TF}{signum_TF}

\item{t_prime}{timepoint}
}
\value{
Sum of the transient part of the RTF and offset p_0 for a defined
time point t_prime.
}
\description{
Calculate sum of the transient part of the RTF and offset p_0
for a defined time point.
}
\examples{
getSignalTransPlusOffset(
        tau_1 = 1.00, tau_2 = 1.00,
        A_trans = 3.05,
        p_0 = -0.28, T_shift = -1, signum_TF = 1,
        t_prime = 0.6)
}
