% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimData.R
\name{getSimData}
\alias{getSimData}
\title{Simulate example data frame for defined RTF parameters}
\usage{
getSimData(modus = "singleDose", noise = NULL, numDoses = 7)
}
\arguments{
\item{modus}{String indicating modus. Either "singleDose" or
"doseDependent" (Default: "singleDose").}

\item{noise}{Noise to be added. Default for modus = "singleDose" is 0.4,
for modus = "doseDependent" 0.02.}

\item{numDoses}{(Only relevant if modus = "doseDependent") Number of distinct
doses to be simulated.}
}
\value{
Data frame with data following an RTF with predefined parameters plus
some added noise.
}
\description{
Simulate an example data frame with data following an RTF with
predefined parameters plus some added noise.
}
\examples{
getSimData()
}
