% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransientFunctionExampleData.R
\name{getTransientFunctionExampleData}
\alias{getTransientFunctionExampleData}
\title{Calculate RTF}
\usage{
getTransientFunctionExampleData(
  tau_1,
  tau_2,
  A_sus,
  A_trans,
  p_0,
  T_shift,
  signum_TF,
  t_prime
)
}
\arguments{
\item{tau_1}{tau_1}

\item{tau_2}{tau_2}

\item{A_sus}{A_sus}

\item{A_trans}{A_trans}

\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{signum_TF}{signum_TF}

\item{t_prime}{timepoint}
}
\value{
Quantitative values resulting from applying RTF
with the specified parameters to a given time point.
}
\description{
Calculate RTF result for a defined time point with defined
parameters
}
\examples{
getTransientFunctionExampleData(
        tau_1 = 1.00, tau_2 = 1.00,
        A_sus = 1.05, A_trans = 3.05,
        p_0 = -0.28, T_shift = -1, signum_TF = 1,
        t_prime = 0.6)
}
