% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillEquation.R
\name{hillEquation}
\alias{hillEquation}
\title{Calculate Hill equation}
\usage{
hillEquation(d, M, h, K)
}
\arguments{
\item{d}{Dose}

\item{M}{Maximum value}

\item{h}{Hill coefficient}

\item{K}{Half-maximum quantity}
}
\value{
Hill equation result value
}
\description{
Calculates Hill equation
}
\examples{
hillEquation(d = 20, M = 5, h = 3, K = 4)
}
