% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillEquationReciprocal.R
\name{hillEquationReciprocal}
\alias{hillEquationReciprocal}
\title{Calculate Hill equation of reciprocal parameter}
\usage{
hillEquationReciprocal(d, M, h, K)
}
\arguments{
\item{d}{Dose}

\item{M}{Maximum value}

\item{h}{Hill coefficient}

\item{K}{Half-maximum quantity}
}
\value{
Hill equation result value
}
\description{
Calculates Hill equation of reciprocal parameter
}
\examples{
hillEquationReciprocal(d = 20, M = 5, h = 3, K = 4)
}
