%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{increaseIndent.RTF}
\alias{increaseIndent.RTF}
\alias{RTF.increaseIndent}
\alias{increaseIndent,RTF-method}

 \title{Increase RTF document indent}

 \description{
	Increase RTF document indent.
 }

 \usage{
\method{increaseIndent}{RTF}(this, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
