% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeOptimObject.R
\name{initializeOptimObject}
\alias{initializeOptimObject}
\title{Initializes OptimObject}
\usage{
initializeOptimObject(data, modus, takeLog10 = TRUE)
}
\arguments{
\item{data}{Data frame containing columns named 't' (time) and 'y'
(quantitative value)}

\item{modus}{String indicating if modus 'RetardedTransientDynamics' or
'ImmediateResponseFunction' should be used}

\item{takeLog10}{Boolean value indicating if log10 of bounds should be applied}
}
\value{
Initialized OptimObject, which is a list containing input data frame
 with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted')
}
\description{
Initializes OptimObject
}
\examples{
data <- getExampleDf()
optimObject <- initializeOptimObject(data, modus = 'RetardedTransientDynamics')
}
