% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowDimensionalRTF.R
\name{lowDimensionalRTF}
\alias{lowDimensionalRTF}
\title{Generate low-dimensional representation of multiple time courses based on
their RTF parameters}
\usage{
lowDimensionalRTF(
  df,
  metaInfo = c(),
  metaInfoName = "Species",
  takeRank = TRUE,
  scaled = TRUE,
  dimX = 1,
  dimY = 2,
  metaInfoSecondRow = FALSE,
  fileString = "lowDimRTF",
  plotItemHeight = 9,
  plotWidth = 22,
  readInParamRdsFilePath = "",
  modelReduction = FALSE,
  nInitialGuesses = 50,
  sameSign = TRUE,
  saveToFile = TRUE,
  numClust = NULL,
  param.data = NULL,
  RTFmodelLst = NULL,
  plotFitsToSingleFile = TRUE,
  plotFitOnly = FALSE,
  plotAllPointsWaterfall = FALSE,
  seed = 111,
  alphaUMAP = 1,
  sizeUMAP = 1.5
)
}
\arguments{
\item{df}{Data frame in which the first column must contain the measurement 
time points, while all subsequent columns correspond to the different time 
courses.}

\item{metaInfo}{String of the column name with meta information (.e.g.
groups.}

\item{metaInfoName}{String of the name of the meta information.}

\item{takeRank}{Boolean indicating if rank should be used for UMAP instead
of absolute value (Default: TRUE)}

\item{scaled}{Boolean indicating if values used for UMAP should be scaled
for the time courses in a cluster (Default: TRUE). Only relevant if
takeRank = FALSE.}

\item{dimX}{Integer from 1 to 3 indicating the UMAP dimension plotted on
the x axis (Default: 1).}

\item{dimY}{Integer from 1 to 3 indicating the UMAP dimension plotted
on the y axis (Default: 2).}

\item{metaInfoSecondRow}{Boolean indicating if second row contains the meta
info.}

\item{fileString}{String the output files should contain.}

\item{plotItemHeight}{Integer indicating height of each plot type in the
output plot.}

\item{plotWidth}{Integer indicating width of output plot.}

\item{readInParamRdsFilePath}{If the RTF parameters to time courses have
already been determined previously, the path of the resulting .rds file can
be specified  using readInParamRdsFilePath, so that the time-intensive
calculation of the RTF parameter does not have to be repreated a second time.}

\item{modelReduction}{Boolean indicating if model reduction should be
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses
(in addition to the default initial guess) used for all four combinations
of signSus and signTrans, which both can be -1 or 1 (Default: 50).}

\item{sameSign}{Boolean indicating if sign of sustained RTF part
(signSus) and transient RTF part (signTrans) should be equal
(Default: TRUE) (Only relevant if param.data is not provided).}

\item{saveToFile}{Boolean indicating if results should be saved to file}

\item{numClust}{(Optional) Number of clusters. If not specified, number of
clusters will be detrmined automatically using the function
NbClust::NbClust().}

\item{param.data}{(Optional) If a data frame with the RTF parameter values
is already available, you can provide it here for saving time.
Its columns 'alpha', 'gamma', 'A', 'B', 'b', 'tau', 'signSus',
and 'signTrans' represent the different RTF parameters. The rows
correspond to the different time courses.}

\item{RTFmodelLst}{(Optional) If a list with the RTF result for each time
course calculated, e.g. using the function getParamsFromMultipleTimeCourses()
or lowDimensionalRTF() is already available, you can provide it here for
saving time.}

\item{plotFitsToSingleFile}{Boolean indicating if plots should be returned
as a single file.}

\item{plotFitOnly}{Plot fit only without additional information as provided
using function plotRTF().}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be
plotted in waterfall plot (Default: FALSE).
If FALSE, all values up to the median of those values are plotted.}

\item{seed}{Seed for UMAP.}

\item{alphaUMAP}{Alpha value (between 0 and 1) of the data points in the
static UMAP plot (Default: 1).}

\item{sizeUMAP}{Size of the data points in the static UMAP plot
(Default: 1.5).}
}
\value{
List with data frame of RTF parameters for each time course
('param.data'), a list with the RTF result for each time course
('RTFmodelLst'), data frame of UMAP results ('umap.data'), data frame of
dynamics per cluster ('dynamics.data'), and ggplot2 plots of the results
('plots'). Function also saves those objects to files if
saveToFile=TRUE.
}
\description{
Generates low-dimensional representation of multiple time
courses based on their RTF parameters.
}
\examples{
\dontrun{
data(strasen)
df <- strasen[, 1:20]
colNames <- colnames(df[2:ncol(df)])
metaInfo <- species <- sub("_[^_]+$", "", colNames)
res <- lowDimensionalRTF(
    df = df,
    metaInfo = metaInfo,
    fileString = "strasen_subset",
    metaInfoName = "Species",
    saveToFile = FALSE
)
ggplot2::ggsave(
    filename = "test.pdf", plot = res[["plots"]],
    width = 10, height = 30
)
}
}
