% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyTimeParamsByFactor.R
\name{multiplyTimeParamsByFactor}
\alias{multiplyTimeParamsByFactor}
\title{Multiply time-dependent variables by a factor.}
\usage{
multiplyTimeParamsByFactor(timeParam, factorVal = NULL, gradientNames = c())
}
\arguments{
\item{timeParam}{Vector of time-dependent variables.}

\item{factorVal}{Numeric value representing factor value the time-dependent
variables should be multiplied with.}

\item{gradientNames}{Names of parameters for which derivatives are
calculated "M","h","K" should be part of it.}
}
\value{
List with vector of transformed time-dependent variables (if
gradientNames are not provided) or matrix of derivatives (if gradientNames
are provided)('timeParam') and factorVal.
}
\description{
Multiply time-dependent variables by a factor.
}
\examples{
t <- c(2, 5, 4, 0.9, 1.2)

# Without gradient
multiplyTimeParamsByFactor(timeParam = t, factorVal = max(t))

# With gradient
multiplyTimeParamsByFactor(
    timeParam = c(tau = 1), factorVal = max(t),
    gradientNames = c("tau", "alpha")
)
}
