%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


\alias{view}
\alias{addTable}
\alias{done}

\alias{addHeader}
\alias{addNewLine}
\alias{addPageBreak}
\alias{addParagraph}
\alias{addPlot}
\alias{addText}
\alias{addTrellisObject}
\alias{decreaseIndent}
\alias{endParagraph}
\alias{increaseIndent}
\alias{setFontSize}
\alias{startParagraph}

\alias{addHeader.RTF}
\alias{addNewLine.RTF}
\alias{addPageBreak.RTF}
\alias{addParagraph.RTF}
\alias{addPlot.RTF}
\alias{addText.RTF}
\alias{addTrellisObject.RTF}
\alias{decreaseIndent.RTF}
\alias{endParagraph.RTF}
\alias{increaseIndent.RTF}
\alias{setFontSize.RTF}
\alias{startParagraph.RTF}



\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have a specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
}

\author{Michael E. Schaffer}


\keyword{documentation}
\keyword{internal}
