% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objFunctGradient.R
\name{objFunctGradient}
\alias{objFunctGradient}
\title{Gradient of log likelihood optimization function}
\usage{
objFunctGradient(par, data, optimObject)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{data}{Data frame containing columns named 't' (time), 'y' (quantitative
value) and optionally 'sdExp' (standard deviation of the experimental data)}

\item{optimObject}{optimObject}
}
\value{
Named vector with optimization value for each parameter
}
\description{
Gradient of log likelihood optimization function
}
