% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimObject.R
\docType{class}
\name{OptimObject-class}
\alias{OptimObject-class}
\title{Class: OptimObject}
\description{
Description for OptimObject
}
\section{Slots}{

\describe{
\item{\code{data}}{Data frame with time resolved data, containing at least the
numeric columns 'y' ( measured outcome) and 't' (time)}

\item{\code{initialGuess.vec}}{Named vector of initial guesses}

\item{\code{lb.vec}}{Named vector of lower bounds}

\item{\code{ub.vec}}{Named vector of upper bounds}

\item{\code{fixed}}{Named vector of fixed parameters}

\item{\code{takeLog10}}{Boolean if log10 is applied to bounds}

\item{\code{positive.par.names}}{Named vector of the parameters having only values
above 0 in initialGuess.vec, lb.vec, and ub.vec}

\item{\code{modus}}{'RetardedTransientDynamics' or 'ImmediateResponseFunction'}

\item{\code{fitted}}{List of values of fitted parameters}
}}

\examples{
OptimObject1 <- new("OptimObject",
  data = data.frame(y=c(0.5, 4, 73), t = c(2, 5, 7)),
  initialGuess.vec = c(par1 = 2, par2 = 54),
  lb.vec = c(par1 = 0, par2 = 0),
  ub.vec = c(par1 = 50, par2 = 50),
  fixed = c(par1 = 4),
  takeLog10 = TRUE,
  positive.par.names = c("par1", "par2"),
  modus = 'RetardedTransientDynamics',
  fitted = c(par1 = 4, par2 = 2)
 )
}
