% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Generate plots}
\usage{
plot(optimObject, modus, plotAllFits = TRUE)
}
\arguments{
\item{modus}{"RetardedTransientDynamics" or 
"DoseDependentRetardedTransientDynamics"}
}
\value{
ggplot object
}
\description{
Plot fitting results for optimObjects of modus 
"RetardedTransientDynamics" or "DoseDependentRetardedTransientDynamics"
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
                data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
optim.res <- getMultiStartResults(
                optimObject.orig, objFunct, nInitialGuesses)
pl <- plotMultiStartPlots(optim.res, fileNamePrefix = "ExamplePlot")
}
