% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Generate plots of data}
\usage{
plotData(data)
}
\arguments{
\item{data}{Data frame with columns t, y, and d (optional)}
}
\value{
ggplot object
}
\description{
Plot data. If there are doses, i.e., if there is a column d,
plot data points of different doses in different colours.
}
\examples{
data.doseResponse <- getExampleDf(modus = "doseDependent")
plotData(data.doseResponse)
}
