% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseResponse.R
\name{plotDoseResponse}
\alias{plotDoseResponse}
\title{Generate plots for dose-response RTF}
\usage{
plotDoseResponse(optimObject, fileNamePrefix = "")
}
\arguments{
\item{optimObject}{optimObject containing element "finalModel"}

\item{fileNamePrefix}{File name prefix. If length>0 plots will be written to
file, otherwise they will be plotted directly.}
}
\value{
ggplot object
}
\description{
Plot fitting results for optimObjects of modus 
"DoseDependentRetardedTransientDynamics"
}
\examples{
data.doseResponse <- getDoseResponseExampleDf()
plot(data.doseResponse)
res.doseResponse <- runRTF(data.doseResponse, 
                          modus = "DoseDependentRetardedTransientDynamics")
plotDoseResponse(res.doseResponse)
}
