% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot RTF together with experimental data points}
\usage{
plotFit(
  par,
  withData = FALSE,
  modus = "RetardedTransientDynamics",
  y = NULL,
  t = NULL,
  d = NULL,
  title = "",
  alphaVal = 0.5
)
}
\arguments{
\item{par}{Vector containing 
alpha, gamma, A, B, b, tau, and signum_TF (modus 
'RetardedTransientDynamics') or 
M_alpha, h_alpha,  K_alpha, M_gamma, h_gamma, K_gamma, 
M_A, h_A, K_A, M_B, h_B, K_B, M_tau, h_tau, K_tau (modus 
'DoseDependentRetardedTransientDynamics').}

\item{withData}{Boolean indicating if data should be added to fit line}

\item{modus}{Modus ('RetardedTransientDynamics' or 
'DoseDependentRetardedTransientDynamics')}

\item{y}{Experimental outcome for time points (corresponds to y column in
experimental data frame)}

\item{t}{Time points}

\item{d}{Dose (obligatory for 'DoseDependentRetardedTransientDynamics')}

\item{title}{Plot title}

\item{alphaVal}{Transparency of points}
}
\value{
ggplot object showing RTF together with experimental data points
}
\description{
Plot RTF together with experimental data points
}
\examples{
gg <- plotFit(
       par = c(alpha = 1.00, gamma = 1.00,
       A = 1.05, B = 3.05,
       b = -0.28, tau = -1, signum_TF = 1),
       withData = TRUE,
       y = c(0.45, 0.96, 1.13, 1.1, 0.9, 0.76, 0.78),
       t = c(0, 0.7, 1.2, 1.55, 2.3, 7.45, 10))
}
