% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLowDimensionalRTF.R
\name{plotLowDimensionalRTF}
\alias{plotLowDimensionalRTF}
\title{Generate plots of low dimensional representation of RTF parameters}
\usage{
plotLowDimensionalRTF(
  df,
  metaInfo,
  metaInfoName,
  takeRank = FALSE,
  scaled = TRUE,
  maxTime = 10,
  numClust = NULL
)
}
\arguments{
\item{df}{Data frame with the RTF parameter values. 
Its columns 'alpha', 'gamma', 'A', 'B', 'b', 'tau', and 'signum_TF'
represent the different RTF parameters. The rows correspond to the different
time series.}

\item{metaInfo}{String of the column name with meta information (.e.g.
groups.}

\item{metaInfoName}{String of the name of the meta information.}

\item{takeRank}{Boolean indicating if rank should be used for UMAP instead 
of absolute value (Default: FALSE)}

\item{scaled}{Boolean indicating if values used for UMAP should be scaled 
for the time series in a cluster (Default: TRUE). Only relevant if 
takeRank = FALSE.}

\item{maxTime}{Time point up to which the dynamics should be plotted.}

\item{numClust}{(Optional) Integer indicating the number of clusters used 
for k-means clustering. If not specified, number of clusters will be 
determined automatically using the function NbClust::NbClust().}
}
\value{
List containing UMAP plot colored by meta info ('umap.metaInfo'), 
UMAP plot colored by cluster ID ('umap.cluster'), plot of parameter 
distribution ('parDistribution'), unscaled dynamics plots 
('dynamics.notscaled'), scaled dynamics plots ('dynamics.scaled'), 
UMAP result data frame ('umap.data'), and data frame with the dynamics values 
('dynamics.data')
}
\description{
Generates plots of low dimensional representation of RTF 
parameters
}
\examples{
data(strasenParams)
metaInfo <- sub("_[^_]+$", "", row.names(strasenParams))
metaInfoName <- "Species"
res <- plotLowDimensionalRTF(df = strasenParams,
                             metaInfo = metaInfo,
                             metaInfoName = metaInfoName,
                             maxTime = 10)
}
