% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiStartPlots.R
\name{plotMultiStartPlots}
\alias{plotMultiStartPlots}
\title{Plot multi start plots}
\usage{
plotMultiStartPlots(optimObject, titlePrefix = "", plotAllFits = TRUE)
}
\arguments{
\item{optimObject}{optimObject, which contains optimResults and bestOptimResult}

\item{titlePrefix}{File prefix}

\item{plotAllFits}{Boolean indicating if all fits should be plotted}
}
\value{
ggplot object
}
\description{
Plot fitting results.
Plots include for each fit plot of the components of the transient function
and the position of the fit in the waterfall plot based on the value of the
optimization measure the fit reached. For the best fit the components of
the transient function are plotted as well as the parameter distribution
over all fits and a water fall plot of all fits.
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
                data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
optim.res <- getMultiStartResults(
                optimObject.orig, objFunct, nInitialGuesses)
pl <- plotMultiStartPlots(optim.res, titlePrefix = "ExamplePlot")
}
