% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNonLinearTransformation.R
\name{plotNonLinearTransformation}
\alias{plotNonLinearTransformation}
\title{Plot sum of nonlinear transformation and offset p_0 together with
experimental data points}
\usage{
plotNonLinearTransformation(p_0, T_shift, y, t_prime)
}
\arguments{
\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{y}{Experimental outcome for time points (corresponds to y column in
experimental data frame)}

\item{t_prime}{timepoints}
}
\value{
ggplot object showing sum of nonlinear transformation and offset p_0
together with experimental data points y.
}
\description{
Plot sum of nonlinear transformation and offset p_0
together with experimental data points y.
}
\examples{
gg <- plotNonLinearTransformation(
       p_0 = -0.28, T_shift = -1,
       y = c(0.45, 0.96, 1.13, 1.1, 0.9, 0.76, 0.78),
       t_prime = c(0, 0.7, 1.2, 1.55, 2.3, 7.45, 10))
}
