% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParamDistribution.R
\name{plotParamDistribution}
\alias{plotParamDistribution}
\title{Plot histograms of fitted parameters}
\usage{
plotParamDistribution(df.parameterValues.long, xLabelSize = 5, yLabelSize = 5)
}
\arguments{
\item{df.parameterValues.long}{Data frame of the parameter values which
have been determined by each of the fits. Two columnns: parameter name
('variable'), value of parameter ('value').}

\item{xLabelSize}{Numeric value of x-axis label sizes.}

\item{yLabelSize}{Numeric value of y-axis label sizes.}
}
\value{
ggplot object of the histograms of the all parameters resulting
from each fit
}
\description{
Plot histograms of fitted parameters
}
\examples{
parameters <- c("par1", "par2", "par3", "par4", "par5")
means <- c(0.5, 2.3, 1.2, 6.5, 4.1)
sds <- c(0.02, 1.2, 0.1, 0.8, 1.1)
nFits <- 50
values <- c()
for (i in 1:length(parameters)) {
    parameter <- parameters[i]
    values <- c(values, rnorm(n = nFits, mean = means[i], sd = sds[i]))
}
df.parameterValues.long <- data.frame(
    variable = rep(parameters, each = nFits),
    value = values
)
gg.paramDistr <- plotParamDistribution(df.parameterValues.long)
}
