% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTF.R
\name{plotRTF}
\alias{plotRTF}
\title{Generate plots of RTF results}
\usage{
plotRTF(
  optimObject,
  fileNamePrefix = "",
  plotTitle = "",
  plotAllFits = TRUE,
  plotAllPointsWaterfall = FALSE,
  pointAlpha = 0.5,
  pointSize = 0.75,
  lineAlpha = NULL,
  lineWidth = NULL,
  color = "#290AD8"
)
}
\arguments{
\item{optimObject}{optimObject containing elements "finalModel" and "modus"}

\item{fileNamePrefix}{File name prefix. If length>0 plots will be written to
file, otherwise they will only be plotted directly. Default is empty string.}

\item{plotTitle}{Title of the plot (Default is no title).}

\item{plotAllFits}{Boolean indicating if all fits should be plotted. Only use
if fileNamePrefix is given. Only relevant for "singleDose"
modus. (Default: TRUE)}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be
plotted in waterfall plot (Default: FALSE).
If FALSE, all values up to the median of those values are plotted.}

\item{pointAlpha}{Transparency of points in plot of fit (Default: 0.5).}

\item{pointSize}{Point size for data points in plot of fit (Default: 0.75).}

\item{lineAlpha}{Transparency of line in plot of fit (Default: 0.5 for 
modus 'singleDose', else 1).}

\item{lineWidth}{Width of line in plot of fit (Default: 1.5 for modus 
'singleDose', else 1).}

\item{color}{Only relevant if modus = 'singleDose': Color of line and points
in plot of fit (Default: "#290AD8")}
}
\value{
ggplot object
}
\description{
Plot RTF results
}
\examples{
data <- getSimData(modus = "singleDose")
res <- RTF(data)
plotRTF(res, plotAllFits = FALSE)
}
