% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTFComponents.R
\name{plotRTFComponents}
\alias{plotRTFComponents}
\title{Plot t against the components of the RTF}
\usage{
plotRTFComponents(pars, data, signum_TF, title = "")
}
\arguments{
\item{pars}{Named vector of the values of the model parameters}

\item{data}{Data frame which needs contain columns 'y' (with quantitative
value of outcome) and 't'}

\item{signum_TF}{Sign of transient function. Positive (1) or negative (-1) sign}

\item{title}{String of the title of the plot}
}
\value{
Combined plot of scatter plot of  t vs.
the components of the retarded transient function
(Nonlinear transformation + p_0, sustained signal + p_0,
transient signal + p_0, retarded transient function (sustained signal +
transient signal + p_0))
}
\description{
Plot t against the components of the retarded transient
function (RTF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
initialGuess.vec.lst <- getInitialGuessVec(
  initialGuess.vec = optimObject.orig$initialGuess.vec,
  lb.vec = optimObject.orig$lb.vec,
  ub.vec = optimObject.orig$ub.vec,
  nInitialGuesses = nInitialGuesses)
initialGuessResults <- runOptimization(initialGuess.vec.lst,
                                      optimObject.orig, objFunct)
res.lst <- initialGuessResults[["res.lst"]]
bestOptimRes <- initialGuessResults[["bestOptimRes"]]
gg.final <- plotRTFComponents(
  pars = bestOptimRes$par,
  data = optimObject.orig$data,
  signum_TF = optimObject.orig$fixed[["signum_TF"]], title = ""
)
}
