% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTFForAllTimeSeries.R
\name{plotRTFForAllTimeSeries}
\alias{plotRTFForAllTimeSeries}
\title{Generate plots based on the list of RTF parameters to multiple time series}
\usage{
plotRTFForAllTimeSeries(res.lst, fileString = "", height = 12, width = 10)
}
\arguments{
\item{res.lst}{List with the RTF result for each time series}

\item{fileString}{String that should be added to file name}

\item{height}{Integer indicating page height}

\item{width}{Integer indicating page width}
}
\value{
pdf file, where each page corresponds to one time series.
}
\description{
Generates plots using plotRTF based on the RTF paramaters 
derived for each of multiple time series.
}
\examples{
data(strasenTimeSeries)
df.multipleTimeSeries <- strasenTimeSeries[, 1:3]
res.lst <- runRTFOnMultipleTimeSeries(df.multipleTimeSeries)
plotRTFForAllTimeSeries(res.lst)
}
