% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignalSus.R
\name{plotSignalSus}
\alias{plotSignalSus}
\title{Plot sum of the sustained part of the RTF and offset p_0 together with
experimental data points}
\usage{
plotSignalSus(tau_1, A_sus, p_0, T_shift, signum_TF, y, t)
}
\arguments{
\item{tau_1}{tau_1}

\item{A_sus}{A_sus}

\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{signum_TF}{signum_TF}

\item{y}{Experimental outcome for time points (corresponds to y column in
experimental data frame)}

\item{t}{Time points}
}
\value{
ggplot object showing sustained part of the RTF and offset p_0
together with experimental data points y.
}
\description{
Plot sum of the sustained part of the RTF and offset p_0
together with experimental data points y.
}
\examples{
gg <- plotSignalSus(
       tau_1 = 1.00,
       A_sus = 1.05,
       p_0 = -0.28, T_shift = -1, signum_TF = 1,
       y = c(0.45, 0.96, 1.13, 1.1, 0.9, 0.76, 0.78),
       t = c(0, 0.7, 1.2, 1.55, 2.3, 7.45, 10))
}
