% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTvsTprime.R
\name{plotTvsTprime}
\alias{plotTvsTprime}
\title{Plot unscaled time plotted against scaled time}
\usage{
plotTvsTprime(t, t_prime)
}
\arguments{
\item{t}{Vector of unscaled time points}

\item{t_prime}{Vector of time points scaled such that each time points lies
in time interval [0,10]}
}
\value{
ggplot object showing unscaled time plotted against scaled time
}
\description{
Plot unscaled time plotted against scaled time.
}
\examples{
t <- c(0.4, 4.5, 5.8, 7.9, 10.02, 15.6, 21.0, 44.8)
t <- t-min(t)
t_range <- max(t)-min(t)
t_prime <- 10*t/t_range
gg <- plotTvsTprime(t, t_prime)
}
