% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Generates UMAP plot based on the RTF parameters for multiple time series}
\usage{
plotUMAP(
  df,
  groupColName = "",
  takeRank = TRUE,
  scaled = TRUE,
  alpha = 0.3,
  size = 0.8,
  ellipse = TRUE,
  ellipseLevel = 0.68,
  seed = 111
)
}
\arguments{
\item{df}{Data frame containing columns corresponding to the values for each
RTF parameter as well as a column indicating how the data points in the UMAP
plot should be colored (name of this column given by 'groupColName').}

\item{groupColName}{String with the column name of the groups based on which 
the data points of the UMAP plot should be colored.}

\item{takeRank}{Boolean indicating if RTF parameters should be scaled
(Default: TRUE)}

\item{scaled}{Boolean indicating if values used for UMAP should be scaled 
for the time series in a cluster (Default: TRUE). Only relevant if 
takeRank = FALSE.}

\item{alpha}{Alpha value (between 0 and 1) of the data points in the UMAP 
plot}

\item{size}{Size of the data points in the UMAP plot}

\item{ellipse}{Boolean indicating if ellipse should be drawn}

\item{ellipseLevel}{The level at which to draw an ellipse (between 0 and 1)}

\item{seed}{Seed}
}
\value{
ggplot2 object of UMAP plot ('plot')
}
\description{
Generates UMAP plot based on the RTF parameters for multiple 
time series
}
\examples{
data(strasenParams)
metaInfo <- sub("_[^_]+$", "", row.names(strasenParams))
metaInfoName <- "Species"
param.df.wMetaInfo <- data.frame(strasenParams, metaInfo)
colnames(param.df.wMetaInfo) <- c(colnames(strasenParams), metaInfoName)
gg.umap.metaInfo <- plotUMAP(df = param.df.wMetaInfo,
                                groupColName = metaInfoName,
                                alpha = 1, size = 1.5)
}
