% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfallPlot.R
\name{plotWaterfallPlot}
\alias{plotWaterfallPlot}
\title{Generate waterfall plot of all fits}
\usage{
plotWaterfallPlot(waterfallValues, idxCurrentFit = NULL)
}
\arguments{
\item{waterfallValues}{List of optimization measures or, if idxCurrentFit is
 given, Dataframe with columns indicating for each fit "value" 
 (optimization measure), "idx" (position of a fit in the sequence
of ordered values), "col" (HEX values for color that should be given to each
 fit)}

\item{idxCurrentFit}{Index indicating the position of the fit in the
waterfall plot that should be highlighted.}
}
\value{
ggplot object of the waterfall plot
}
\description{
Generate waterfall plot of all fits. Optionally a specific fit
can be highlighted using the argument 'idxCurrentFit'.
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
                        data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
res.lst.wFinal <- getMultiStartResults(optimObject.orig, objFunct,
                    nInitialGuesses)

optimResults <- res.lst.wFinal$optimResults

optimResults.optimRes <- lapply(optimResults, function(x) {
  tmp <- unlist(x[grep("optimRes", names(x))], recursive = FALSE)
  names(tmp) <- sub("optimRes.", "", names(tmp))
  tmp
})

optimResTmpLstValuesAll <- unlist(
  lapply(optimResults.optimRes, function(x) unlist(x[grep("value",names(x))])))
gg <- plotWaterfallPlot(optimResTmpLstValuesAll)
}
