% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDoseDependent-data.R
\docType{data}
\name{resDoseDependent}
\alias{resDoseDependent}
\title{Example RTF parameter data frame to simulated time courses with dose 
information}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(resDoseDependent)
}
\description{
Example RTF() output for simulated data frame with time-resolved data 
and dose information
 
\preformatted{
data.doseDependent <- getSimData(modus = "doseDependent")
resDoseDependent <- RTF(data.doseDependent, modus = "doseDependent")
}
}
\examples{
data(resDoseDependent)
}
\keyword{datasets}
