% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resSingleDose-data.R
\docType{data}
\name{resSingleDose}
\alias{resSingleDose}
\title{Example RTF parameter data frame to simulated time courses}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(resSingleDose)
}
\description{
Example RTF() output for simulated data frame with time-resolved data
 
\preformatted{
data.singleDose <- getSimData(modus = "singleDose")
resSingleDose <- RTF(data.singleDose, modus = "singleDose")
}
}
\examples{
data(resSingleDose)
}
\keyword{datasets}
