% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTFOnMultipleTimeCourses.R
\name{RTFOnMultipleTimeCourses}
\alias{RTFOnMultipleTimeCourses}
\title{Run RTF on multiple time courses}
\usage{
RTFOnMultipleTimeCourses(
  df,
  modelReduction = FALSE,
  nInitialGuesses = 50,
  sameSign = TRUE
)
}
\arguments{
\item{df}{Data frame in which the first column must contain the measurement 
time points, while all subsequent columns correspond to the different time 
courses.}

\item{modelReduction}{Boolean indicating if model reduction should be
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses
(in addition to the default initial guess) used for all four combinations
of signSus and signTrans, which both can be -1 or 1 (Default: 50).}

\item{sameSign}{Boolean indicating if sign of sustained RTF part
(signSus) and transient RTF part (signTrans) should be equal
(Default: TRUE).}
}
\value{
List with the RTF result for each time course.
}
\description{
Runs RTF on multiple columns corresponding to different time
courses.
}
\examples{
data(strasen)
df <- strasen[, 1:3]
RTFmodelLst <- RTFOnMultipleTimeCourses(df)
}
