% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTF.R
\name{runRTF}
\alias{runRTF}
\title{Run RTF}
\usage{
runRTF(
  data,
  modus = "RetardedTransientDynamics",
  optimFunction = "chiSquare",
  control = list(trace = 1, maxit = 1000, factr = 1e-20)
)
}
\arguments{
\item{data}{Data frame containing columns named 't' (time) and
'y' (quantitative value)}

\item{modus}{String indicating if modus 'RetardedTransientDynamics' or
'ImmediateResponseFunction' should be used}

\item{optimFunction}{String indicating the optimization function which 
should be used (Default: "chiSquare")}

\item{control}{List of control arguments passed to the function stats::optim 
(Default: list(trace = 1, maxit = 1000, factr = 1.0e-20))}
}
\value{
List of the final RTF model (finalModel), the optimized parameters
(finalParams), the plot of the final model (finalPlot), as well as the
intermediate results (intermediateResults).
}
\description{
Run RTF
}
\examples{
modus <- "RetardedTransientDynamics"
data <- getExampleDf()
plot(data)
res <- runRTF(data, modus = modus)
}
