% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTFOnMultipleTimeSeries.R
\name{runRTFOnMultipleTimeSeries}
\alias{runRTFOnMultipleTimeSeries}
\title{Run RTF on multiple times series}
\usage{
runRTFOnMultipleTimeSeries(df, modelReduction = FALSE, nInitialGuesses = 100)
}
\arguments{
\item{df}{Data frame with the first column named 'time'
defining the different time points and all the following columns 
corresponding to the different time series.}

\item{modelReduction}{Boolean indicating if model reduction should be 
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 100).}
}
\value{
List with the RTF result for each time series
}
\description{
Runs RTF on multiple columns corresponding to different times 
series.
}
\examples{
data(strasenTimeSeries)
df <- strasenTimeSeries[, 1:3]
res.lst <- runRTFOnMultipleTimeSeries(df)
}
