% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFittedRTF.R
\name{saveFittedRTF}
\alias{saveFittedRTF}
\title{Saves the function of the fitted RTF as well as a table with the fitted 
parameters and the respective bounds to files}
\usage{
saveFittedRTF(optimObject, functionName = "fittedRTF")
}
\arguments{
\item{optimObject}{optimObject, containing the fitted parameters and the 
modus ('singleDose' or 'doseDependent').}

\item{functionName}{(Optional) 
String, which specifies how the fitted RTF function in "[functionName].R"  
and the table with the fitted parameters and respective bounds 
("[functionName]_fittedParams.tsv") should be called.
If not specified, it will be set to "fittedRTF".}
}
\value{
Saves the function of the fitted RTF as well as the fitted parameters
to files.
}
\description{
Saves the function of the fitted RTF (to .R file) as well as a 
table with the fitted parameters and the respective bounds (to .tsv file). 
After the function is generated it becomes sourced.
}
\examples{
\dontrun{
data(resSingleDose)
saveFittedRTF(optimObject = resSingleDose$finalModel)
}
}
