% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTimeCol.R
\name{scaleTimeCol}
\alias{scaleTimeCol}
\title{Add new column with data points scaled to lie between 0 and 10}
\usage{
scaleTimeCol(data, scaleTime = TRUE)
}
\arguments{
\item{data}{Data frame which contains time resolved data,
with columns 't' (time) and 'y' (quantitative value)}

\item{scaleTime}{Boolean value indicating if time should be scaled should be
scaled such that they lie in the interval [0, 10]}
}
\value{
Data frame which contains an additional column 't_prime'
}
\description{
Adds column 't_prime' to input data frame, which scaleTime is
set to TRUE corresponds to 't' scaled to interval [0, 10],
otherwise 't_prime' corresponds to 't'
}
\examples{
data <- getExampleDf()
data.scaled <- scaleTimeCol(data, scaleTime = TRUE)
}
