% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTimeParameter.R
\name{scaleTimeParameter}
\alias{scaleTimeParameter}
\title{Get vectored with data points scaled to lie between 0 and 10}
\usage{
scaleTimeParameter(timeParam, maxVal = NULL, reverse = FALSE)
}
\arguments{
\item{timeParam}{Time-dependent variable}

\item{maxVal}{Maximum of range the time-dependent variable should be scaled to}

\item{reverse}{Boolean if scaling should be reversed}
}
\value{
Scaled vector if reverse = FALSE and unscaled timeParam is entered,
unscaled vector if reverse = TRUE and already scaled timeParam is entered.
}
\description{
Scales or reverses scaling for time-dependent variable
}
\examples{
t <- c(2, 5, 4, 0.9, 1.2)
t_prime <- scaleTimeParameter(timeParam=t, maxVal = max(t))
}
