% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSmallerModelIfDiffIsSmall.R
\name{selectSmallerModelIfDiffIsSmall}
\alias{selectSmallerModelIfDiffIsSmall}
\title{Select smaller model if there is no significant difference to bigger model}
\usage{
selectSmallerModelIfDiffIsSmall(res, res.smallerModel)
}
\arguments{
\item{res}{optimObject for which no paramater of the retarded transient
has been fixed prior to fitting.}

\item{res.smallerModel}{optimObject for which parameters are fitted prior to
fitting.}
}
\value{
optimObject of the smaller model if for the bigger and smaller model
there is no significance difference in the likelihood values achieved 
through fitting.
An optimObject is a list containing input data frame
with time resolved data ('data'), the vector of initial guesses 
('initialGuess.vec'), of lower bounds ('lb.vec'), of upper bounds ('ub.vec'),
vector of fixed parameters ('fixed'), if log10 is applied to bounds 
('takeLog10'), the parameters having no negative values in initialGuess.vec, 
lb.vec, and ub.vec ('positive.par.names'), modus ('modus'), 
likelihood function for the parameter optimization ('optimFunction'), 
list of control parameters passed to stats::optim ('control'),
and a list of values of fitted parameters ('fitted', can be empty).
}
\description{
Model reduction: Select smaller model if the difference in the
likelihood value is not significant for Chi-square significance test.
}
\examples{
data <- getSimData()
modus <- "timeDependent"
optimObject.orig <- initializeOptimObject(data, modus = modus)
res.all <- getInitialGuessResults(optimObject.orig)
optimObjectTmp <- optimObject.orig
optimObjectTmp$positive.par.names <-
                      setdiff(optimObjectTmp$positive.par.names, "tau")
optimObjectTmp$fixed[["tau"]] <- optimObject.orig$lb.vec[["tau"]]
res.tauLB <- getInitialGuessResults(optimObjectTmp)
res <- selectSmallerModelIfDiffIsSmall(res.all, res.tauLB)
}
