%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setFontSize.RTF}
\alias{setFontSize.RTF}
\alias{RTF.setFontSize}
\alias{setFontSize.RTF}
\alias{setFontSize,RTF-method}


 \title{Set RTF document font size}

 \description{
	Set RTF document font size.
 }

 \usage{\method{setFontSize}{RTF}(this, font.size, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{font.size}{New default font size in points.}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }
\keyword{internal}
\keyword{methods}
