%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{startParagraph.RTF}
\alias{startParagraph.RTF}
\alias{RTF.startParagraph}
\alias{startParagraph.RTF}
\alias{startParagraph,RTF-method}


 \title{Start a new paragraph in the RTF document}

 \description{
	Start a new paragraph in the RTF document.
 }

 \usage{\method{startParagraph}{RTF}(this, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }
\keyword{internal}
\keyword{methods}
