% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeLog10OfBounds.R
\name{takeLog10OfBounds}
\alias{takeLog10OfBounds}
\title{Log10-transform parameters}
\usage{
takeLog10OfBounds(optimObject)
}
\arguments{
\item{optimObject}{OptimObject, which is a list containing input data frame
 with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted')}
}
\value{
OptimObject with  values of vector of initial guesses
('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec') of OptimObject being log10 transformed.
}
\description{
Replace the parameter values of vector of initial guesses
('initialGuess.vec'), of lower bounds ('lb.vec'),
of upper bounds ('ub.vec') of OptimObject with the log10 transformed values.
Only those parameter values are replaced which have values bigger than 0 in
all three vectors, which needs to be indicated by the respective parameter
name being included in the vector 'positive.par.names' of the OptimObject.
}
\examples{
data <- getExampleDf()
optimObject <- initializeOptimObject(data,
                                    modus = 'RetardedTransientDynamics',
                                    takeLog10=FALSE)
positive.par.names <- getPositiveParNames(lb.vec = optimObject$lb.vec,
                                          ub.vec = optimObject$ub.vec,
                                          initialGuess.vec =
                                                optimObject$initialGuess.vec)
positive.par.names <- setdiff(positive.par.names, "sigma")
optimObject$positive.par.names <- positive.par.names
optimObject.log10 <- takeLog10OfBounds(optimObject)
}
