%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{view.RTF}
\alias{view.RTF}
\alias{RTF.view}
\alias{view,RTF-method}

 \title{View encoded RTF}

 \description{
	View encoded RTF.
 }

 \usage{
\method{view}{RTF}(this, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{Not used.}
 }

 \value{
 	Output the content of the object as RTF code.
 }

 \author{Michael Schaffer}

 \seealso{
 	For more information see \code{\link{RTF}}.
 }


\keyword{methods}
