%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addHeader.RTF}
\alias{addHeader.RTF}
\alias{RTF.addHeader}
\alias{addHeader,RTF-method}

 \title{Insert a header into the RTF document}

 \description{
	Insert a header into the RTF document.
 }

 \usage{
\method{addHeader}{RTF}(this, title, subtitle=NULL, font.size=NULL, TOC.level=NULL, ...)
}

 \arguments{
 	\item{this}{An RTF object.}
   \item{title}{Header title text.}
   \item{subtitle}{Header subtitle text. \bold{optional}.}
	\item{font.size}{Font size in points. \bold{optional}.}
	\item{TOC.level}{Indent level for table of contents. \bold{optional}.}
 	\item{...}{Not used.}
 }

 \author{Michael Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }


\keyword{methods}
