%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rtf.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addText.RTF}
\alias{addText.RTF}
\alias{RTF.addText}
\alias{addText.RTF}
\alias{addText,RTF-method}


 \title{Insert text into the RTF document}

 \description{
	Insert text into the RTF document.
 }

 \usage{\method{addText}{RTF}(this, ...)}

 \arguments{
 	\item{this}{An RTF object.}
 	\item{...}{A character \code{\link[base]{vector}} of text to add.}
 }

 \author{Michael E. Schaffer}

 \seealso{
   For more information see \code{\link{RTF}}.
 }
\keyword{internal}
\keyword{methods}
