% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFittingResult.R
\name{getBestFittingResult}
\alias{getBestFittingResult}
\title{Fit function for positive OR negative sign}
\usage{
getBestFittingResult(
  optimObject,
  parStr = "signum_TF",
  parVal = 1,
  nInitialGuesses = 100
)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame
with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of
lower bounds ('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and
ub.vec ('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted').}

\item{parStr}{String of variable based on which comparison is conducted (Default: "signum_TF")}

\item{parVal}{Value of variable defined in parString  (Default: 1)}

\item{nInitialGuesses}{Integer corresponding to number of initial guesses 
(Default: 100)}
}
\value{
optimObject, to which fitting result of the best fit ('fitted'),
optimization measure value of this fit ('value').
}
\description{
Fits function for positive OR negative sign (signum_TF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
res.all.plus <- getBestFittingResult(
        optimObject.orig, parStr = "signum_TF", parVal = 1)
}
