% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleDf.R
\name{getExampleDf}
\alias{getExampleDf}
\title{Generate example data frame for defined RTF parameters}
\usage{
getExampleDf(modus = "timeDependent")
}
\value{
Data frame with data following an RTF with predefined parameters plus
some added noise.
}
\description{
Generates an example dataframe with data following an RTF with
predefined parameters plus some added noise.
}
\examples{
getExampleDf()
}
