% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFittingResult.R
\name{getFittingResult}
\alias{getFittingResult}
\title{Fit function for positive AND negative sign}
\usage{
getFittingResult(
  optimObject,
  parStr = "signum_TF",
  parVals = c(-1, 1),
  nInitialGuesses = 50
)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame
with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds
('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec
('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted')}

\item{parStr}{Parameter for which different parameter values
parVals are used for the fitting (Default: "signum_TF")}

\item{parVals}{Vector of different values of parameter parStr used for
fitting (Default: c(-1, 1))}

\item{nInitialGuesses}{Integer corresponding to number of initial guesses 
(Default: 50)}
}
\value{
List of two optimObjects for a positive and a negative signum_TF
('plus1' and 'minus1'). Each including fitting result of the best
fit ('fitted'), optimization measure value of this fit ('value').
}
\description{
Fits function for positive AND negative sign (signum_TF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(
            data, modus = 'RetardedTransientDynamics')
res.all.plusMinus <- getFittingResult(
            optimObject.orig, 
            parStr = "signum_TF", parVals = c(-1, 1),
            titlePrefixPrefix = "fullModel_")
}
