% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialGuessResults.R
\name{getInitialGuessResults}
\alias{getInitialGuessResults}
\title{Fit function for initial guesses}
\usage{
getInitialGuessResults(optimObject, nInitialGuesses = 100)
}
\arguments{
\item{optimObject}{optimObject, which is a list containing input data frame
with time resolved data ('data'),
the vector of initial guesses ('initialGuess.vec'), of lower bounds
('lb.vec'),
of upper bounds ('ub.vec'), vector of fixed parameters ('fixed'),
if log10 is applied to bounds ('takeLog10'), the parameters having no
negative values in initialGuess.vec, lb.vec, and ub.vec
('positive.par.names'),
modus ('modus'), and a list of values of fitted parameters ('fitted')}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 100).}
}
\value{
optimObject including fitting result of the best
fit ('fitted'), optimization measure value of this fit ('value').
}
\description{
Fits function for initial guesses
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data, modus = 'timeDependent')
res.all <- getInitialGuessResults( optimObject.orig)
}
