% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonLinTransformationPlusOffset.R
\name{getNonLinTransformationPlusOffset}
\alias{getNonLinTransformationPlusOffset}
\title{Calculate sum of nonlinear transformation and offset p_0}
\usage{
getNonLinTransformationPlusOffset(p_0, T_shift, t_prime)
}
\arguments{
\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{t_prime}{timepoint}
}
\value{
Sum of nonlinear transformation and offset p_0 for a defined
time point t_prime.
}
\description{
Calculate sum of nonlinear transformation and offset p_0
for a defined time point.
}
\examples{
getNonLinTransformationPlusOffset(
        p_0 = -0.28,
        T_shift = -1,
        t_prime = 0.6)
}
