% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonLinTransformationPlusOffset.R
\name{getNonLinTransformationPlusOffset}
\alias{getNonLinTransformationPlusOffset}
\title{Calculate sum of nonlinear transformation and offset b}
\usage{
getNonLinTransformationPlusOffset(b, tau, t)
}
\arguments{
\item{b}{b}

\item{tau}{tau}

\item{t}{Time point}
}
\value{
Sum of nonlinear transformation and offset b for a defined
time point t.
}
\description{
Calculate sum of nonlinear transformation and offset b
for a defined time point.
}
\examples{
getNonLinTransformationPlusOffset(
        b = -0.28,
        tau = -1,
        t = 0.6)
}
