% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamDf.R
\name{getParamDf}
\alias{getParamDf}
\title{Generate data frame containing the RTF parameters for each time series.}
\usage{
getParamDf(res.lst)
}
\arguments{
\item{res.lst}{List with the RTF result for each time series}
}
\value{
Data frame containing the RTF parameters for each time 
series, where the rows correspond to the different time series and 
the columns to the RTF parameters.
}
\description{
Generate data frame of RTF parameters for each time series from
list of RTF results.
}
\examples{
data(strasenTimeSeries)
df.multipleTimeSeries <- strasenTimeSeries[, 1:3]
res.lst <- RTFOnMultipleTimeSeries(df.multipleTimeSeries)
param.df <- getParamDf(res.lst)
}
