% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamsFromMultipleTimeSeries.R
\name{getParamsFromMultipleTimeSeries}
\alias{getParamsFromMultipleTimeSeries}
\title{Get RTF parameters for multiple time series}
\usage{
getParamsFromMultipleTimeSeries(
  df,
  doRTFPlots = TRUE,
  fileString = "lowDimRTF",
  readInParamRdsFilePath = "",
  saveFolderPath = "",
  modelReduction = FALSE,
  nInitialGuesses = 100,
  plotFitsToSingleFile = TRUE,
  plotFitOnly = FALSE,
  plotAllPointsWaterfall = FALSE
)
}
\arguments{
\item{df}{Data frame with the first column named 'time' defining the
different time points and all the following columns corresponding to the
different time series.}

\item{doRTFPlots}{Boolean indicating if RTF plots should be generated.}

\item{fileString}{String the name of the created .rds file of the RTF
parameters should contain. If this string has length 0 or if
readInParamRdsFilePath is defined no .rds file will be created.}

\item{readInParamRdsFilePath}{if the RTF parameters to time series have
already been determined previously, the path of the resulting .rds file can
be specified  using readInParamRdsFilePath, so that the time-intensive
calculation of the RTF parameter does not have to be repreated a second time.}

\item{saveFolderPath}{Path of folder to where plots and rds files should be 
saved. Default: current folder.}

\item{modelReduction}{Boolean indicating if model reduction should be
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 100).}

\item{plotFitsToSingleFile}{Boolean indicating if plots should be returned as a
single file.}

\item{plotFitOnly}{Plot fit only without additional information as provided
using function plotRTF().}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
Data frame containing the RTF parameters for each time
series, where the rows correspond to the different time series and
the columns to the RTF parameters.
}
\description{
Get RTF parameters for multiple time series specified in
data frame.
}
\examples{
data(strasenTimeSeries)
df <- strasenTimeSeries[, 1:3]
param.df <- getParamsFromMultipleTimeSeries(df)
}
