% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignalSusPlusOffset.R
\name{getSignalSusPlusOffset}
\alias{getSignalSusPlusOffset}
\title{Calculate sum of the sustained part of the RTF and offset p_0}
\usage{
getSignalSusPlusOffset(tau_1, A_sus, p_0, T_shift, signum_TF, t_prime)
}
\arguments{
\item{tau_1}{tau_1}

\item{A_sus}{A_sus}

\item{p_0}{p_0}

\item{T_shift}{T_shift}

\item{signum_TF}{signum_TF}

\item{t_prime}{timepoint}

\item{tau_2}{tau_2}
}
\value{
Sum of the sustained part of the RTF and offset p_0 for a defined
time point t_prime.
}
\description{
Calculate sum of the sustained part of the RTF and offset p_0
for a defined time point.
}
\examples{
getSignalSusPlusOffset(
        tau_1 = 1.00, tau_2 = 1.00,
        A_sus = 1.05,
        p_0 = -0.28, T_shift = -1, signum_TF = 1,
        t_prime = 0.6)
}
