% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignalTransPlusOffset.R
\name{getSignalTransPlusOffset}
\alias{getSignalTransPlusOffset}
\title{Calculate sum of the transient part of the RTF and offset b}
\usage{
getSignalTransPlusOffset(alpha, gamma, B, b, tau, signum_TF, t)
}
\arguments{
\item{alpha}{alpha}

\item{gamma}{gamma}

\item{B}{B}

\item{b}{b}

\item{tau}{tau}

\item{signum_TF}{signum_TF}

\item{t}{Time point}
}
\value{
Sum of the transient part of the RTF and offset b for a defined
time point t.
}
\description{
Calculate sum of the transient part of the RTF and offset b
for a defined time point.
}
\examples{
getSignalTransPlusOffset(
        alpha = 1.00, gamma = 1.00,
        B = 3.05,
        b = -0.28, tau = -1, signum_TF = 1,
        t = 0.6)
}
