% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimData.R
\name{getSimData}
\alias{getSimData}
\title{Generate example data frame for defined RTF parameters}
\usage{
getSimData(modus = "timeDependent")
}
\arguments{
\item{modus}{String indicating modus. Either "timeDependent" or 
"doseDependent". Default: "timeDependent"}
}
\value{
Data frame with data following an RTF with predefined parameters plus
some added noise.
}
\description{
Generates an example dataframe with data following an RTF with
predefined parameters plus some added noise.
}
\examples{
getSimData()
}
