% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransientFunctionResult.R
\name{getTransientFunctionResult}
\alias{getTransientFunctionResult}
\title{Calculate RTF results for multiple time points}
\usage{
getTransientFunctionResult(
  rtfPar = c(),
  t = NULL,
  signum_TF = 1,
  scale = TRUE,
  calcGradient = FALSE
)
}
\arguments{
\item{rtfPar}{Named vector of the parameter values used for RTF.}

\item{t}{Vector of time points}

\item{signum_TF}{Value of signum_TF (-1 or 1, Default: 1)}

\item{scale}{Boolean, indicates if time-dependent parameters t, tau,
alpha, and gamma should be scaled}

\item{calcGradient}{Boolean indicating if gradient should be calculated
(Default: FALSE)}
}
\value{
Vector of predicted quantitative values resulting from applying RTF
with the specified parameters to time points given in t.
}
\description{
Apply RTF with defined parameters for defined time points
}
\examples{
rtfPar <- c(alpha = 1.00, gamma = 1.00, A = 1.05,
         B = 3.05, b = -0.28, tau = -1)
t <- c(0, 0.71, 1.42, 2.14, 2.85, 3.57, 4.28, 5, 5.71, 6.42,
             7.14, 7.85, 8.57, 9.28, 10)
y <- getTransientFunctionResult(rtfPar = rtfPar, t = t, signum_TF = 1)
plot(t, y)
}
