% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillEquation.R
\name{hillEquation}
\alias{hillEquation}
\title{Calculate Hill equation}
\usage{
hillEquation(
  d,
  M,
  h,
  K,
  reciprocal = FALSE,
  calcGradient = FALSE,
  minval = NULL
)
}
\arguments{
\item{d}{Dose}

\item{M}{Maximum value}

\item{h}{Hill coefficient}

\item{K}{Half-maximum quantity}

\item{reciprocal}{Boolean indicating if Hill equation is calculated for
reciprocal parameter}

\item{calcGradient}{Boolean indicating if gradient should be calculated}

\item{minval}{(optional) Numeric, defines the lowest value the result can 
acquire. Only relevant if reciprocal = TRUE.}
}
\value{
Hill equation result value
}
\description{
Calculates Hill equation
}
\examples{
hillEquation(d = 20, M = 5, h = 3, K = 4)
}
