% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsumoto-data.R
\docType{data}
\name{matsumoto}
\alias{matsumoto}
\title{Experimental time series data with dose dependencies}
\format{
An object of class \code{data.frame} with 5280 rows and 6 columns.
}
\usage{
data(matsumoto)
}
\description{
Time series data of amino acids showing dose dependencies for the 
ingestion of branched-chain amino acids, taken from the publication of 
Matsumoto et al. 2014. (https://doi.org/10.1186/2193-1801-3-35).
}
\examples{
data(matsumoto)
}
\references{
Matsumoto, T., Nakamura, K., Matsumoto, H. et al. Bolus 
ingestion of individual branched-chain amino acids alters plasma amino acid 
profiles in young healthy men. SpringerPlus 3, 35 (2014).
(https://doi.org/10.1186/2193-1801-3-35)
}
\keyword{datasets}
