% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelReduction.R
\name{modelReduction}
\alias{modelReduction}
\title{Run model reduction on full RTF model}
\usage{
modelReduction(
  res,
  nInitialGuesses = 50,
  plotAllPointsWaterfall = FALSE,
  saveParams = TRUE,
  saveParamsFileString = "fittedParams"
)
}
\arguments{
\item{res}{Full RTF model, i.e., generated without fixed parameters}

\item{nInitialGuesses}{Integer indicating number of initial guesses
(in addition to the default initial guess) used for all four combinations
of signSus and signTrans, which both can be -1 or 1 (Default: 50).}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be
plotted in waterfall plot (Default: FALSE).
If FALSE, all values up to the median of those values are plotted.}

\item{saveParams}{(Optional) Boolean, indicating if the fitted parameters 
should be saved to files (Default: TRUE).}

\item{saveParamsFileString}{(Optional, only relevant if saveParams = TRUE) 
String, which specifies how the output files with the fitted parameters 
generated by saveFinalParams() should be called ("[saveParamsFileString].tsv"
and "[saveParamsFileString].R"). If not specified it will be set to 
"fittedParams".}
}
\value{
List of RTF model after model reduction ('finalModel') and
corresponding model parameters ('finalParams'). Additionally the fitted 
parameters will be saved to files.
}
\description{
Run model reduction on full RTF model
}
\examples{
# Single-dose RTF
modus <- "singleDose"
data <- getSimData(modus = modus)
plotData(data)
res <- RTF(data, modus = modus)
res.reduced <- modelReduction(res$finalModel)

\dontrun{
# Dose-dependent RTF
modus <- "doseDependent"
data.dose <- getSimData(modus = modus)
plotData(data.dose)
res.dose <- RTF(data.dose, modus = modus)
res.reduced.dose <- modelReduction(res.dose$finalModel)
}
}
