% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelReduction.R
\name{modelReduction}
\alias{modelReduction}
\title{Run model reduction on full RTF model}
\usage{
modelReduction(res, nInitialGuesses = 100, plotAllPointsWaterfall = FALSE)
}
\arguments{
\item{res}{Full RTF model, i.e., generated without fixed parameters}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 100).}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
RTF model after model reduction
}
\description{
Run model reduction on full RTF model
}
\examples{
modus <- "timeDependent"
data <- getExampleDf()
plotData(data)
res <- runRTF(data, modus = modus)
res.reduced <- modelReduction(res$finalModel)
}
