% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelReduction.R
\name{modelReduction}
\alias{modelReduction}
\title{Run model reduction on full RTF model}
\usage{
modelReduction(res, nInitialGuesses = 50, plotAllPointsWaterfall = FALSE)
}
\arguments{
\item{res}{Full RTF model, i.e., generated without fixed parameters}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 50).}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
RTF model after model reduction for time-dependent RTF, and a data
frame indicating dose-dependency of each RTF parameter for dose-dependent 
RTF.
}
\description{
Run model reduction on full RTF model
}
\examples{
# Time-dependent RTF
modus <- "timeDependent"
data <- getSimData(modus = modus)
plotData(data)
res <- RTF(data, modus = modus)
res.reduced <- modelReduction(res$finalModel)

\dontrun{
# Dose-dependent RTF
modus <- "doseDependent"
data.dose <- getSimData(modus = modus)
plotData(data.dose)
res.dose <- RTF(data.dose, modus = modus)
df.dose <- modelReduction(res.dose$finalModel)
}
}
