%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


\alias{view}
\alias{addTable}
\alias{done}
\alias{addPageBreak}
\alias{startParagraph}
\alias{endParagraph}
\alias{addHeader}
\alias{addParagraph}
\alias{addText}
\alias{addNewLine}
\alias{decreaseIndent}
\alias{increaseIndent}
\alias{setFontSize}
\alias{addPlot}
\alias{addPng}
\alias{addTrellisObject}
\alias{addTOC}
\alias{addSessionInfo}

\alias{rtf.forest.plot.xscale}


\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have a specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
}

\author{Michael Schaffer}


\keyword{documentation}
\keyword{internal}
