% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objFunctGradient.R
\name{objFunctGradient}
\alias{objFunctGradient}
\title{Gradient of log likelihood optimization function}
\usage{
objFunctGradient(par, data, optimObject, calcGradient)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{data}{Data frame containing columns named 't' (time), 'y' (quantitative
value) and optionally 'sigmaExp' (standard error of the experimental data)}

\item{optimObject}{optimObject}

\item{calcGradient}{Placeholder so that stats::optim is not returning  the
error 'Error in gr(par, ...) : unused argument (calcGradient = FALSE)'}
}
\value{
Named vector with optimization value for each parameter
}
\description{
Gradient of log likelihood optimization function
}
