% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot RTF together with experimental data points}
\usage{
plotFit(
  par,
  withData = FALSE,
  modus = "RetardedTransientDynamics",
  y = NULL,
  t = NULL,
  d = NULL,
  plotType = "all",
  title = "",
  alpha = 0.5
)
}
\arguments{
\item{par}{Vector containing 
tau_1, tau_2, A_sus, A_trans, p_0, T_shift, and signum_TF (modus 
'RetardedTransientDynamics') or 
M_tau1, h_tau1,  K_tau1, M_tau2, h_tau2, K_tau2, M_Asus, h_Asus, K_Asus, 
M_Atrans, h_Atrans, K_Atrans, M_Tshift, h_Tshift, K_Tshift (modus 
'DoseDependentRetardedTransientDynamics').}

\item{withData}{Boolean indicating if data should be added to fit line}

\item{modus}{Modus ('RetardedTransientDynamics' or 
DoseDependentRetardedTransientDynamics')}

\item{y}{Experimental outcome for time points (corresponds to y column in
experimental data frame)}

\item{t}{Time points}

\item{d}{Dose (obligatory for 'DoseDependentRetardedTransientDynamics')}

\item{plotType}{String indication if transient function should be plotted
('all') (Default) or if in case of modus = 'RetardedTransientDynamics'
 components of transient function ("nonLinearTransformationOnly", 
"transientOnly", "sustainedOnly") should be plotted.}

\item{title}{Plot title}

\item{alpha}{Transparency of points}
}
\value{
ggplot object showing RTF together with experimental data points
}
\description{
Plot RTF together with experimental data points
}
\examples{
gg <- plotFit(
       par = c(tau_1 = 1.00, tau_2 = 1.00,
       A_sus = 1.05, A_trans = 3.05,
       p_0 = -0.28, T_shift = -1, signum_TF = 1),
       withData = TRUE,
       y = c(0.45, 0.96, 1.13, 1.1, 0.9, 0.76, 0.78),
       t = c(0, 0.7, 1.2, 1.55, 2.3, 7.45, 10))
}
