% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot RTF, optionally together with experimental data points}
\usage{
plotFit(
  par,
  withData = FALSE,
  modus = "timeDependent",
  y = NULL,
  t = NULL,
  d = NULL,
  title = "",
  pointAlpha = 0.5,
  lineAlpha = 0.5,
  pointSize = 0.75
)
}
\arguments{
\item{par}{Vector containing
alpha, gamma, A, B, b, tau, and signum_TF (modus
'timeDependent') or
M_alpha, h_alpha,  K_alpha, M_gamma, h_gamma, K_gamma,
M_A, h_A, K_A, M_B, h_B, K_B, M_tau, h_tau, K_tau (modus
'doseDependent').}

\item{withData}{Boolean indicating if data should be added to fit line}

\item{modus}{Modus ('timeDependent' or
'doseDependent')}

\item{y}{Experimental outcome for time points (corresponds to y column in
experimental data frame)}

\item{t}{Time points}

\item{d}{Dose (obligatory for 'doseDependent')}

\item{title}{Plot title}

\item{pointAlpha}{Transparency of points}

\item{lineAlpha}{For modus 'timeDependent': Transparency of lines}

\item{pointSize}{Point size}
}
\value{
ggplot object showing RTF, if withData = TRUE together with 
experimental data points
}
\description{
Plot RTF, optionally together with experimental data points. 
In addition, the sustained and transient part of the RTF are plotted.
}
\examples{
gg <- plotFit(
       par = c(alpha = 1.00, gamma = 1.00,
       A = 1.05, B = 3.05,
       b = -0.28, tau = -1, signum_TF = 1),
       withData = TRUE,
       y = c(0.45, 0.96, 1.13, 1.1, 0.9, 0.76, 0.78),
       t = c(0, 0.7, 1.2, 1.55, 2.3, 7.45, 10))
}
