% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLowDimensionalRTF.R
\name{plotLowDimensionalRTF}
\alias{plotLowDimensionalRTF}
\title{Generate plots of low-dimensional representation of multiple time series, 
which is based on the RTF parameters of those time series.}
\usage{
plotLowDimensionalRTF(
  df,
  metaInfo,
  metaInfoName,
  takeRank = FALSE,
  scaled = TRUE,
  dimX = 1,
  dimY = 2,
  maxTime = 10,
  numClust = NULL,
  seed = 111,
  alphaUMAP = 1,
  sizeUMAP = 1.5,
  doPlotInteractiveUMAP = FALSE,
  timeSeriesDf = NULL,
  RTFmodelLst = NULL,
  conditions = c(),
  species = c(),
  plotLines = TRUE,
  hRatio = 0,
  vRatio = 0
)
}
\arguments{
\item{df}{Data frame with the RTF parameter values. 
Its columns 'alpha', 'gamma', 'A', 'B', 'b', 'tau', and 'signum_TF'
represent the different RTF parameters. The rows correspond to the different
time series.}

\item{metaInfo}{String of the column name with meta information (.e.g.
groups.}

\item{metaInfoName}{String of the name of the meta information.}

\item{takeRank}{Boolean indicating if rank should be used for UMAP instead 
of absolute value (Default: FALSE)}

\item{scaled}{Boolean indicating if values used for UMAP should be scaled 
for the time series in a cluster (Default: TRUE). Only relevant if 
takeRank = FALSE.}

\item{dimX}{Integer from 1 to 3 indicating the UMAP dimension plotted on 
the x axis (Default: 1).}

\item{dimY}{Integer from 1 to 3 indicating the UMAP dimension plotted 
on the y axis (Default: 2).}

\item{maxTime}{Time point up to which the dynamics should be plotted.}

\item{numClust}{(Optional) Integer indicating the number of clusters used 
for k-means clustering. If not specified, number of clusters will be 
determined automatically using the function NbClust::NbClust().}

\item{seed}{Seed for UMAP.}

\item{alphaUMAP}{Alpha value (between 0 and 1) of the data points in the 
static UMAP plot (Default: 1).}

\item{sizeUMAP}{Size of the data points in the static UMAP plot 
(Default: 1.5).}

\item{doPlotInteractiveUMAP}{Boolean indicating if interactive UMAP should be
generated (Default: FALSE).}

\item{timeSeriesDf}{(Only relevant if doPlotInteractiveUMAP = TRUE) 
Data frame with the first column corresponding to the time points 
and all the following columns corresponding to the different time series.}

\item{RTFmodelLst}{(Only relevant if doPlotInteractiveUMAP = TRUE) 
List with the RTF result for each time series calculated using the function 
getParamsFromMultipleTimeSeries().}

\item{conditions}{(Only relevant if doPlotInteractiveUMAP = TRUE) 
Vector specifying the condition for each time series. If plotLines = TRUE,
number of different conditions may not exceed 2.}

\item{species}{(Only relevant if doPlotInteractiveUMAP = TRUE) 
Vector with names of molecular species for each time series. Needs to be 
provided in plotLines = TRUE.}

\item{plotLines}{(Only relevant if doPlotInteractiveUMAP = TRUE) 
Boolean indicating if lines should be plotted between species of the same 
name (Default: TRUE). Can only be plotted if there are at most two time 
series per molecular species.}

\item{hRatio}{(Only relevant if doPlotInteractiveUMAP = TRUE)
Float between 0 and 1 indicating where subplot should be placed
horizontally in relation to plot width (Default: 0).}

\item{vRatio}{(Only relevant if doPlotInteractiveUMAP = TRUE)
Float between 0 and 1 indicating where subplot should be placed
vertically in relation to plot height (Default: 0).}
}
\value{
List containing UMAP plot colored by meta info ('umap.metaInfo'), 
UMAP plot colored by cluster ID ('umap.cluster'), interactive UMAP plot 
colored by condition ('umap.interactive'), plot of parameter 
distribution ('parDistribution'), unscaled dynamics plots 
('dynamics.notscaled'), scaled dynamics plots ('dynamics.scaled'), 
UMAP result data frame ('umap.data'), and data frame with the dynamics values 
('dynamics.data')
}
\description{
Generates plots of low-dimensional representation of multiple 
time series, which is based on the RTF parameters of those time series.
}
\examples{
data(strasenParams)
metaInfo <- sub("_[^_]+$", "", row.names(strasenParams))
metaInfoName <- "Species"
res <- plotLowDimensionalRTF(df = strasenParams,
                             metaInfo = metaInfo,
                             metaInfoName = metaInfoName,
                             maxTime = 10,
                             doPlotInteractiveUMAP = FALSE)
}
