% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTF.R
\name{plotRTF}
\alias{plotRTF}
\title{Generate plots}
\usage{
plotRTF(
  optimObject,
  fileNamePrefix = "",
  plotTitle = "",
  plotAllFits = TRUE,
  plotAllPointsWaterfall = FALSE
)
}
\arguments{
\item{optimObject}{optimObject containing elements "finalModel" and "modus"}

\item{fileNamePrefix}{File name prefix. If length>0 plots will be written to
file, otherwise they will be plotted directly. Default is empty string.}

\item{plotTitle}{Title of the plot (Default is no title).}

\item{plotAllFits}{Boolean indicating if all fits should be plotted. Only use
if fileNamePrefix is given. Only relevant for "RetardedTransientDynamics"
modus. (Default: TRUE)
using function plotRTF().}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
ggplot object
}
\description{
Plot fitting results for optimObjects of modus
"RetardedTransientDynamics" or "DoseDependentRetardedTransientDynamics"
}
\examples{
data.doseResponse <- getExampleDf(
                            modus = "DoseDependentRetardedTransientDynamics")
plotData(data.doseResponse)
res.doseResponse <- runRTF(data.doseResponse, modelReduction = FALSE)
plotRTF(res.doseResponse, plotAllFits = FALSE)
}
