% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTF.R
\name{plotRTF}
\alias{plotRTF}
\title{Generate plots of RTF results}
\usage{
plotRTF(
  optimObject,
  fileNamePrefix = "",
  plotTitle = "",
  plotAllFits = TRUE,
  plotAllPointsWaterfall = FALSE
)
}
\arguments{
\item{optimObject}{optimObject containing elements "finalModel" and "modus"}

\item{fileNamePrefix}{File name prefix. If length>0 plots will be written to
file, otherwise they will be plotted directly. Default is empty string.}

\item{plotTitle}{Title of the plot (Default is no title).}

\item{plotAllFits}{Boolean indicating if all fits should be plotted. Only use
if fileNamePrefix is given. Only relevant for "timeDependent"
modus. (Default: TRUE)}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
ggplot object
}
\description{
Plot RTF results for optimObjects of modus
"timeDependent" or "doseDependent"
}
\examples{
data <- getExampleDf(modus = "timeDependent")
res <- runRTF(data)
plotRTF(res, plotAllFits = FALSE)
}
