% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTFComponents.R
\name{plotRTFComponents}
\alias{plotRTFComponents}
\title{Plot t against the components of the RTF}
\usage{
plotRTFComponents(pars, data, title = "")
}
\arguments{
\item{pars}{Named vector of the values of the model parameters}

\item{data}{Data frame which needs contain columns 'y' (with quantitative
value of outcome) and 't'}

\item{title}{String of the title of the plot}
}
\value{
Combination of scatter plot of t vs.
the components of the retarded transient function
(Nonlinear transformation + b, sustained signal + b,
transient signal + b, retarded transient function (sustained signal +
transient signal + b))
}
\description{
Plot t against the components of the retarded transient
function (RTF)
}
\examples{
data <- getExampleDf()
optimObject.orig <- initializeOptimObject(data,
                                         modus = 'RetardedTransientDynamics')
optimObject.orig$fixed[["signum_TF"]] <- 1
nInitialGuesses <- 50
initialGuess.vec.lst <- getInitialGuessVec(
  initialGuess.vec = optimObject.orig$initialGuess.vec,
  lb.vec = optimObject.orig$lb.vec,
  ub.vec = optimObject.orig$ub.vec,
  nInitialGuesses = nInitialGuesses)
initialGuessResults <- runOptimization(initialGuess.vec.lst,
                                      optimObject.orig, objFunct)
res.lst <- initialGuessResults[["res.lst"]]
bestOptimRes <- initialGuessResults[["bestOptimResult"]]
gg.final <- plotRTFComponents(
  pars = bestOptimRes$par,
  data = optimObject.orig$data,
  title = ""
)
}
