% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRTFForAllTimeSeries.R
\name{plotRTFForAllTimeSeries}
\alias{plotRTFForAllTimeSeries}
\title{Generate plots based on the list of RTF parameters to multiple time series}
\usage{
plotRTFForAllTimeSeries(
  res.lst,
  fileString = "",
  height = 12,
  width = 10,
  plotFitsToSingleFile = TRUE,
  plotFitsToSingleFileExtension = "jpeg",
  plotFitOnly = FALSE,
  plotAllPointsWaterfall = FALSE
)
}
\arguments{
\item{res.lst}{List with the RTF result for each time series}

\item{fileString}{String that should be added to file name}

\item{height}{Integer indicating page height}

\item{width}{Integer indicating page width}

\item{plotFitsToSingleFile}{Boolean indicating if plots should be returned as a
single file.}

\item{plotFitsToSingleFileExtension}{Image file extension, if 
plotFitsToSingleFile=TRUE (Default: jpeg.)}

\item{plotFitOnly}{Plot fit only without additional information as provided 
using function plotRTF().}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}
}
\value{
pdf file, where each page corresponds to one time series.
}
\description{
Generates plots using plotRTF based on the RTF paramaters 
derived for each of multiple time series.
}
\examples{
data(strasenTimeSeries)
df.multipleTimeSeries <- strasenTimeSeries[, 1:3]
res.lst <- runRTFOnMultipleTimeSeries(df.multipleTimeSeries)
plotRTFForAllTimeSeries(res.lst)
}
