% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfallPlot.R
\name{plotWaterfallPlot}
\alias{plotWaterfallPlot}
\title{Generate waterfall plot of all fits}
\usage{
plotWaterfallPlot(waterfallValues, idxCurrentFit = NULL, plotAllPoints = FALSE)
}
\arguments{
\item{waterfallValues}{List of likelihood values or, if idxCurrentFit is
 given, data frame with columns indicating for each fit "value" 
 (likelihood value), "idx" (position of a fit in the sequence
of ordered values), "col" (HEX values for color that should be given to each
 fit)}

\item{idxCurrentFit}{Index indicating the position of the fit in the
waterfall plot that should be highlighted.}

\item{plotAllPoints}{Boolean indicating if all points should be plotted 
(Default: FALSE). If FALSE, all values up to the median of those values 
are plotted.}
}
\value{
ggplot object of the waterfall plot
}
\description{
Generate waterfall plot of all fits. Optionally, a specific fit
can be highlighted using the argument 'idxCurrentFit'.
}
\examples{
data <- getSimData()
optimObject.orig <- initializeOptimObject(data, modus = 'singleDose')
RTFmodelLst.wFinal <- getInitialGuessResults(optimObject.orig,
                                             nInitialGuesses = 50)
optimResults <- RTFmodelLst.wFinal$optimResults
optimResTmpLstValuesAll <- unlist(lapply(optimResults, function(x) {
  unlist(x[grep("value", names(x))])
}))
gg <- plotWaterfallPlot(optimResTmpLstValuesAll)
}
