% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterfallPlot.R
\name{plotWaterfallPlot}
\alias{plotWaterfallPlot}
\title{Generate waterfall plot of all fits}
\usage{
plotWaterfallPlot(optimResTmpLstValuesAll, idxCurrentFit = NULL)
}
\arguments{
\item{idxCurrentFit}{Index indicating the position of the fit in the
waterfall plot that should be highlighted.}

\item{optimObject}{Dataframe with columns indicating for each fit
"value" (optimization measure), "idx" (position of a fit in the sequence
of ordered values), "col" (HEX values for color that should be given to each
 fit)}
}
\value{
ggplot object of the waterfall plot
}
\description{
Generate waterfall plot of all fits. Optionally a specific fit
can be highlighted using the argument 'idxCurrentFit'.
}
\examples{
data <- getExampleDf()
data <- scaleTimeCol(data)
optimObject.orig <- initializeOptimObject(
                        data, modus = 'RetardedTransientDynamics')
signum_TF <- 1
optimObject.orig$fixed[["signum_TF"]] <- signum_TF
nInitialGuesses <- 50
res.lst.wFinal <- getInitialGuessResults(optimObject.orig, objFunct,
                    nInitialGuesses, plot = TRUE)
final <- res.lst.wFinal$final
waterfallPlotData <- final$gg.waterfall$data
plotWaterfallPlot(waterfallPlotData, idxCurrentFit = 4)
}
