% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTFOnMultipleTimeCourses.R
\name{RTFOnMultipleTimeCourses}
\alias{RTFOnMultipleTimeCourses}
\title{Run RTF on multiple time courses}
\usage{
RTFOnMultipleTimeCourses(df, modelReduction = FALSE, nInitialGuesses = 50)
}
\arguments{
\item{df}{Data frame with the first column corresponding to the time points 
and all the following columns corresponding to the different time courses.}

\item{modelReduction}{Boolean indicating if model reduction should be 
performed for RTF}

\item{nInitialGuesses}{Integer indicating number of initial guesses 
(in addition to the default initial guess) used both for a signum_TF of -1 
and 1 (Default: 50).}
}
\value{
List with the RTF result for each time course.
}
\description{
Runs RTF on multiple columns corresponding to different time 
courses.
}
\examples{
data(strasen)
df <- strasen[, 1:3]
RTFmodelLst <- RTFOnMultipleTimeCourses(df)
}
