% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTF.R
\name{runRTF}
\alias{runRTF}
\title{Run RTF}
\usage{
runRTF(data, modus = "RetardedTransientDynamics", plot = TRUE)
}
\arguments{
\item{data}{Data frame containing columns named 't' (time) and
'y' (quantitative value)}

\item{modus}{String indicating if modus 'RetardedTransientDynamics' or
'ImmediateResponseFunction' should be used}

\item{plot}{Boolean value indicating if fitting results should be plotted
and saved as file}
}
\value{
List of the final RTF model (finalModel), the optimized parameters
(finalParams), the plot of the final model (finalPlot), as well as the
intermediate results (intermediateResults).
}
\description{
Run RTF
}
\examples{
modus <- "RetardedTransientDynamics"
plot <- TRUE
data <- getExampleDf()
plot(data)
res <- runRTF(data, modus = modus, plot = plot)
}
