% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRTF.R
\name{runRTF}
\alias{runRTF}
\title{Run RTF}
\usage{
runRTF(
  data,
  modus = NULL,
  optimFunction = "logLikelihood",
  modelReduction = TRUE,
  nInitialGuesses = 100,
  plotAllPointsWaterfall = FALSE,
  control = list(trace = 1, maxit = 1000, factr = 1e+07)
)
}
\arguments{
\item{data}{Data frame containing columns named 't' (time) and
'y' (quantitative value) for modus = 'timeDependent' ('TD'), and columns 
't', 'y', and 'd' (dose) for modus = 'doseDependent' ('DD')}

\item{modus}{String indicating if modus 'timeDependent' ('TD') or
'doseDependent' ('DD') should be used.
If no modus is provided default setting are
'doseDependent' if column with name 'd' is present
and else 'timeDependent'.}

\item{optimFunction}{String indicating the optimization function which
should be used (Default: "logLikelihood")}

\item{modelReduction}{Boolean indicating of model reduction should be
performed (Default: TRUE)}

\item{nInitialGuesses}{Integer indicating number of initial guesses
(Default: 100)}

\item{plotAllPointsWaterfall}{Boolean indicating if all points should be 
plotted in waterfall plot (Default: FALSE). 
If FALSE, all values up to the median of those values are plotted.}

\item{control}{List of control arguments passed to the function stats::optim
(Default: list(trace = 1, maxit = 1000, factr = 1e7))}
}
\value{
List of the final RTF model (finalModel), the optimized parameters
(finalParams), the plot of the final model (finalPlot), as well as the
intermediate results (intermediateResults).
}
\description{
Run RTF
}
\examples{
modus <- "timeDependent"
data <- getExampleDf()
plotData(data)
res <- runRTF(data, modus = modus)
}
