% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFinalParams.R
\name{saveFinalParams}
\alias{saveFinalParams}
\title{Saves fitted parameters to files}
\usage{
saveFinalParams(optimObject, fileString = "fittedParams")
}
\arguments{
\item{optimObject}{optimObject, containing the fitted parameters}

\item{fileString}{(Optional, only relevant if saveParams = TRUE) 
String, which specifies how the output files with the fitted parameters 
should be called ("[fileString].tsv" and "[fileString].R"). 
If not specified it will be set to "fittedParams".}
}
\value{
Saves fitted parameters to files.
}
\description{
Saves the fitted parameters to an R file which becomes sourced 
to make the fitted parameters readily available. Furthermore, the fitted 
parameters together with the respective parameter bounds become saved to a 
.tsv file.
}
\examples{
\dontrun{
data(resSingleDose)
saveFinalParams(optimObject = resSingleDose$finalModel)
}
}
