% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortListByValue.R
\name{sortListByValue}
\alias{sortListByValue}
\title{Sort sublists by log-likelihood value}
\usage{
sortListByValue(lst)
}
\arguments{
\item{lst}{List with sublists, each of which needs to contain
parameter 'value', according to which the sublists will be sorted.}
}
\value{
List sorted according to the parameter 'value' in ascending order.
}
\description{
Sorts a list of sublists in ascending order according to
parameter 'value'.
}
\examples{
lst <- list(
    optimRes = list(str = "res1", value = -20),
    optimRes = list(str = "res2", value = -45),
    optimRes = list(str = "res3", value = 8)
)
lst.sorted <- sortListByValue(lst)
}
